Refactor machine learning models into dedicated modules
