+++ models/ml_models.py
@@ -1,1856 +1,27 @@
-import pandas as pd
-import numpy as np
-import os
-from typing import Dict, List, Tuple, Optional, Any
-from datetime import datetime
 import logging
-import joblib
-from pathlib import Path
+from datetime import datetime
+from typing import Any, Dict, List, Optional, Tuple
+
+import numpy as np
+import pandas as pd
 import torch
 import torch.nn as nn
 import torch.optim as optim
-import xgboost as xgb
-import lightgbm as lgb
-from sklearn.model_selection import train_test_split, cross_val_score
-from sklearn.preprocessing import StandardScaler
-from sklearn.metrics import (
-    mean_squared_error,
-    accuracy_score,
-    mean_absolute_error,
-    r2_score,
-)
+
 from data.stock_data import StockDataProvider
 
+from .cache import AdvancedCacheManager, RedisCache
+from .deep_learning import DeepLearningPredictor, DQNReinforcementLearner
+from .ensemble_predictor import (
+    AdvancedEnsemblePredictor,
+    EnsembleStockPredictor,
+    ParallelStockPredictor,
+)
+from .meta_learning import MetaLearningOptimizer
+from .ml_stock_predictor import MLStockPredictor
+from .sentiment import MacroEconomicDataProvider, SentimentAnalyzer
+
 logger = logging.getLogger(__name__)
-
-
-class MLStockPredictor:
-    """機械学習を使用した株価予測モデル"""
-
-    def __init__(self, model_type: str = "xgboost"):
-        self.model_type = model_type
-        self.data_provider = StockDataProvider()
-        self.scaler = StandardScaler()
-        self.model = None
-        self.is_trained = False
-        self.feature_names = []
-        self.model_path = Path("models/saved_models")
-        self.model_path.mkdir(exist_ok=True)
-
-    def prepare_features(self, data: pd.DataFrame) -> pd.DataFrame:
-        """高度な特徴量を準備する"""
-        if data.empty:
-            return pd.DataFrame()
-        # 技術指標を計算
-        data = self.data_provider.calculate_technical_indicators(data)
-        features = pd.DataFrame(index=data.index)
-        # === 基本価格特徴量 ===
-        features["price_change"] = data["Close"].pct_change()
-        features["volume_change"] = data["Volume"].pct_change()
-        features["high_low_ratio"] = (data["High"] - data["Low"]) / data["Close"]
-        features["close_open_ratio"] = (data["Close"] - data["Open"]) / data["Open"]
-        # === 移動平均関連（強化） ===
-        features["sma_20_ratio"] = data["Close"] / data["SMA_20"]
-        features["sma_50_ratio"] = data["Close"] / data["SMA_50"]
-        features["sma_cross"] = (data["SMA_20"] > data["SMA_50"]).astype(int)
-        features["sma_distance"] = (data["SMA_20"] - data["SMA_50"]) / data["Close"]
-        # 指数移動平均
-        ema_12 = data["Close"].ewm(span=12).mean()
-        ema_26 = data["Close"].ewm(span=26).mean()
-        features["ema_12_ratio"] = data["Close"] / ema_12
-        features["ema_26_ratio"] = data["Close"] / ema_26
-        features["ema_cross"] = (ema_12 > ema_26).astype(int)
-        # === 高度な技術指標 ===
-        features["rsi"] = data["RSI"]
-        features["rsi_normalized"] = (data["RSI"] - 50) / 50  # -1 to 1
-        features["rsi_overbought"] = (data["RSI"] > 70).astype(int)
-        features["rsi_oversold"] = (data["RSI"] < 30).astype(int)
-        features["macd"] = data["MACD"]
-        features["macd_signal"] = data["MACD_Signal"]
-        features["macd_histogram"] = data["MACD"] - data["MACD_Signal"]
-        features["macd_bullish"] = (features["macd_histogram"] > 0).astype(int)
-        features["atr_ratio"] = data["ATR"] / data["Close"]
-        # === ボリンジャーバンド ===
-        bb_middle = data["Close"].rolling(20).mean()
-        bb_std = data["Close"].rolling(20).std()
-        bb_upper = bb_middle + (bb_std * 2)
-        bb_lower = bb_middle - (bb_std * 2)
-        # ゼロ除算を防ぐ安全チェック
-        bb_range = bb_upper - bb_lower
-        features["bb_position"] = (data["Close"] - bb_lower) / bb_range.where(
-            bb_range != 0, 1
-        )
-        features["bb_squeeze"] = bb_range / bb_middle.where(bb_middle != 0, 1)
-        features["bb_breakout_up"] = (data["Close"] > bb_upper).astype(int)
-        features["bb_breakout_down"] = (data["Close"] < bb_lower).astype(int)
-        # === ストキャスティクス ===
-        low_14 = data["Low"].rolling(14).min()
-        high_14 = data["High"].rolling(14).max()
-        features["stoch_k"] = 100 * (data["Close"] - low_14) / (high_14 - low_14)
-        features["stoch_d"] = features["stoch_k"].rolling(3).mean()
-        # === ウィリアムズ%R ===
-        features["williams_r"] = -100 * (high_14 - data["Close"]) / (high_14 - low_14)
-        # === CCIとROC ===
-        typical_price = (data["High"] + data["Low"] + data["Close"]) / 3
-        sma_tp = typical_price.rolling(20).mean()
-        mad = typical_price.rolling(20).apply(lambda x: abs(x - x.mean()).mean())
-        features["cci"] = (typical_price - sma_tp) / (0.015 * mad)
-        features["roc_5"] = (data["Close"] / data["Close"].shift(5) - 1) * 100
-        features["roc_10"] = (data["Close"] / data["Close"].shift(10) - 1) * 100
-        # === ラグ特徴量（強化） ===
-        for lag in [1, 2, 3, 5, 10, 20]:
-            features[f"close_lag_{lag}"] = data["Close"].shift(lag)
-            features[f"volume_lag_{lag}"] = data["Volume"].shift(lag)
-            features[f"rsi_lag_{lag}"] = data["RSI"].shift(lag)
-            features[f"price_change_lag_{lag}"] = data["Close"].pct_change().shift(lag)
-        # === ローリング統計（拡張） ===
-        for window in [3, 5, 10, 20, 50]:
-            features[f"close_mean_{window}"] = data["Close"].rolling(window).mean()
-            features[f"close_std_{window}"] = data["Close"].rolling(window).std()
-            features[f"close_min_{window}"] = data["Close"].rolling(window).min()
-            features[f"close_max_{window}"] = data["Close"].rolling(window).max()
-            features[f"volume_mean_{window}"] = data["Volume"].rolling(window).mean()
-            features[f"volume_std_{window}"] = data["Volume"].rolling(window).std()
-        # === ボラティリティ特徴量 ===
-        for window in [5, 10, 20, 50]:
-            features[f"volatility_{window}d"] = data["Close"].rolling(window).std()
-            features[f"volatility_ratio_{window}d"] = (
-                features[f"volatility_{window}d"] / data["Close"]
-            )
-        # === トレンド強度とモメンタム ===
-        for window in [5, 10, 20, 50]:
-            returns = data["Close"].pct_change()
-            features[f"trend_strength_{window}"] = returns.rolling(window).mean()
-            features[f"momentum_{window}"] = (
-                data["Close"] / data["Close"].shift(window) - 1
-            )
-        # === リターン特徴量 ===
-        for period in [1, 2, 3, 5, 10, 15, 20]:
-            features[f"return_{period}d"] = data["Close"].pct_change(period)
-        # === 高度な統計特徴量 ===
-        for window in [10, 20]:
-            returns = data["Close"].pct_change()
-            features[f"skewness_{window}"] = returns.rolling(window).skew()
-            features[f"kurtosis_{window}"] = returns.rolling(window).kurt()
-        # === ボリューム特徴量（強化） ===
-        features["volume_price_trend"] = (
-            (data["Close"] - data["Close"].shift()) / data["Close"].shift()
-        ) * data["Volume"]
-        features["on_balance_volume"] = (
-            data["Volume"]
-            * ((data["Close"] > data["Close"].shift()).astype(int) * 2 - 1)
-        ).cumsum()
-        features["volume_rate_change"] = data["Volume"].pct_change()
-        # ボリューム移動平均
-        for window in [5, 10, 20]:
-            vol_ma = data["Volume"].rolling(window).mean()
-            features[f"volume_ma_ratio_{window}"] = data["Volume"] / vol_ma
-        # === 季節性・曜日効果 ===
-        features["day_of_week"] = data.index.dayofweek
-        features["day_of_month"] = data.index.day
-        features["month"] = data.index.month
-        features["quarter"] = data.index.quarter
-        # === サポート・レジスタンス指標 ===
-        for window in [10, 20, 50]:
-            rolling_max = data["High"].rolling(window).max()
-            rolling_min = data["Low"].rolling(window).min()
-            features[f"resistance_distance_{window}"] = (
-                rolling_max - data["Close"]
-            ) / data["Close"]
-            features[f"support_distance_{window}"] = (
-                data["Close"] - rolling_min
-            ) / data["Close"]
-        # === 価格パターン指標 ===
-        # ドージ判定
-        body_size = abs(data["Close"] - data["Open"])
-        candle_range = data["High"] - data["Low"]
-        features["doji"] = (body_size / candle_range < 0.1).astype(int)
-        # ハンマー・倒立ハンマー
-        lower_shadow = data["Close"].combine(data["Open"], min) - data["Low"]
-        upper_shadow = data["High"] - data["Close"].combine(data["Open"], max)
-        features["hammer"] = (
-            (lower_shadow > 2 * body_size) & (upper_shadow < body_size)
-        ).astype(int)
-        features["hanging_man"] = (
-            (upper_shadow > 2 * body_size) & (lower_shadow < body_size)
-        ).astype(int)
-        # === 相対パフォーマンス ===
-        # 市場全体との相関を後で計算するプレースホルダー
-        features["market_relative_strength"] = 0  # 後で実装
-        # === ギャップ検出 ===
-        prev_close = data["Close"].shift(1)
-        features["gap_up"] = ((data["Open"] > prev_close * 1.02)).astype(int)
-        features["gap_down"] = ((data["Open"] < prev_close * 0.98)).astype(int)
-        features["gap_size"] = (data["Open"] - prev_close) / prev_close
-        # 欠損値処理
-        features = features.ffill().fillna(0)
-        return features
-
-    def create_targets(
-        self, data: pd.DataFrame, prediction_days: int = 5
-    ) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        """予測ターゲットを作成（分類と回帰の両方）"""
-        targets_regression = pd.DataFrame(index=data.index)
-        targets_classification = pd.DataFrame(index=data.index)
-        # 回帰ターゲット: 将来の価格変化率
-        for days in [1, 3, 5, 10]:
-            future_return = data["Close"].shift(-days) / data["Close"] - 1
-            targets_regression[f"return_{days}d"] = future_return
-        # 分類ターゲット: 価格上昇/下降
-        for days in [1, 3, 5, 10]:
-            future_return = data["Close"].shift(-days) / data["Close"] - 1
-            targets_classification[f"direction_{days}d"] = (future_return > 0).astype(
-                int
-            )
-        # 推奨スコアターゲット（0-100）
-        targets_regression["recommendation_score"] = (
-            self._calculate_future_performance_score(data)
-        )
-        return targets_regression, targets_classification
-
-    def _calculate_future_performance_score(self, data: pd.DataFrame) -> pd.Series:
-        """将来のパフォーマンスに基づくスコアを計算"""
-        scores = pd.Series(index=data.index, dtype=float)
-        for i in range(len(data) - 10):
-            current_price = data["Close"].iloc[i]
-            future_prices = data["Close"].iloc[i + 1 : i + 11]
-            if len(future_prices) < 10:
-                continue
-            # 最大利益
-            max_gain = (future_prices.max() - current_price) / current_price
-            # 最大損失
-            max_loss = (future_prices.min() - current_price) / current_price
-            # 最終リターン
-            final_return = (future_prices.iloc[-1] - current_price) / current_price
-            # スコア計算
-            score = 50  # ベーススコア
-            score += max_gain * 100  # 最大利益を加算
-            score += max_loss * 100  # 最大損失を減算（負の値）
-            score += final_return * 50  # 最終リターンの影響
-            # ボラティリティペナルティ
-            volatility = future_prices.std() / current_price
-            score -= volatility * 20
-            scores.iloc[i] = max(0, min(100, score))
-        return scores.fillna(50)
-
-    def prepare_dataset(
-        self, symbols: List[str], start_date: str = "2020-01-01"
-    ) -> Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:
-        """複数銘柄のデータセットを準備"""
-        all_features = []
-        all_targets_reg = []
-        all_targets_cls = []
-        for symbol in symbols:
-            try:
-                logger.info(f"Processing {symbol}...")
-                data = self.data_provider.get_stock_data(symbol, "3y")
-                if data.empty or len(data) < 100:
-                    logger.warning(f"Insufficient data for {symbol}")
-                    continue
-                features = self.prepare_features(data)
-                targets_reg, targets_cls = self.create_targets(data)
-                # 銘柄情報を追加
-                features["symbol_" + symbol] = 1
-                all_features.append(features)
-                all_targets_reg.append(targets_reg)
-                all_targets_cls.append(targets_cls)
-            except Exception as e:
-                logger.error(f"Error processing {symbol}: {str(e)}")
-                continue
-        if not all_features:
-            raise ValueError("No valid data available")
-        # データを結合
-        combined_features = pd.concat(all_features, ignore_index=True)
-        combined_targets_reg = pd.concat(all_targets_reg, ignore_index=True)
-        combined_targets_cls = pd.concat(all_targets_cls, ignore_index=True)
-        # ワンホットエンコーディング用の銘柄列を調整
-        symbol_columns = [
-            col for col in combined_features.columns if col.startswith("symbol_")
-        ]
-        for symbol in symbols:
-            col_name = f"symbol_{symbol}"
-            if col_name not in combined_features.columns:
-                combined_features[col_name] = 0
-        combined_features = combined_features.fillna(0)
-        return combined_features, combined_targets_reg, combined_targets_cls
-
-    def train_model(
-        self, symbols: List[str], target_column: str = "recommendation_score"
-    ):
-        """モデルを訓練する"""
-        from config.settings import get_settings
-
-        settings = get_settings()
-        logger.info("Preparing dataset...")
-        features, targets_reg, targets_cls = self.prepare_dataset(symbols)
-        # ターゲットが存在するかチェック
-        if (
-            target_column not in targets_reg.columns
-            and target_column not in targets_cls.columns
-        ):
-            raise ValueError(f"Target column {target_column} not found")
-        # ターゲットデータを選択
-        if target_column in targets_reg.columns:
-            targets = targets_reg[target_column]
-            task_type = "regression"
-        else:
-            targets = targets_cls[target_column]
-            task_type = "classification"
-        # 欠損値を除去
-        valid_indices = ~(targets.isna() | features.isna().any(axis=1))
-        features_clean = features[valid_indices]
-        targets_clean = targets[valid_indices]
-        if len(features_clean) < settings.model.min_training_data:
-            raise ValueError(
-                f"Insufficient training data: {len(features_clean)} < {settings.model.min_training_data}"
-            )
-        # 特徴量名を保存
-        self.feature_names = features_clean.columns.tolist()
-        # 時系列分割で訓練・テスト分割
-        train_size = int(len(features_clean) * settings.model.train_test_split)
-        X_train = features_clean.iloc[:train_size]
-        X_test = features_clean.iloc[train_size:]
-        y_train = targets_clean.iloc[:train_size]
-        y_test = targets_clean.iloc[train_size:]
-        # 特徴量スケーリング
-        X_train_scaled = self.scaler.fit_transform(X_train)
-        X_test_scaled = self.scaler.transform(X_test)
-        logger.info(f"Training {self.model_type} model...")
-        logger.info(f"Training data shape: {X_train_scaled.shape}")
-        # モデル訓練
-        if self.model_type == "xgboost":
-            if task_type == "regression":
-                self.model = xgb.XGBRegressor(
-                    n_estimators=settings.model.xgb_n_estimators,
-                    max_depth=settings.model.xgb_max_depth,
-                    learning_rate=settings.model.xgb_learning_rate,
-                    random_state=settings.model.xgb_random_state,
-                    n_jobs=-1,
-                )
-            else:
-                self.model = xgb.XGBClassifier(
-                    n_estimators=settings.model.xgb_n_estimators,
-                    max_depth=settings.model.xgb_max_depth,
-                    learning_rate=settings.model.xgb_learning_rate,
-                    random_state=settings.model.xgb_random_state,
-                    n_jobs=-1,
-                )
-        elif self.model_type == "lightgbm":
-            if task_type == "regression":
-                self.model = lgb.LGBMRegressor(
-                    n_estimators=settings.model.lgb_n_estimators,
-                    max_depth=settings.model.lgb_max_depth,
-                    learning_rate=settings.model.lgb_learning_rate,
-                    random_state=settings.model.lgb_random_state,
-                    n_jobs=-1,
-                    verbose=-1,
-                )
-            else:
-                self.model = lgb.LGBMClassifier(
-                    n_estimators=settings.model.lgb_n_estimators,
-                    max_depth=settings.model.lgb_max_depth,
-                    learning_rate=settings.model.lgb_learning_rate,
-                    random_state=settings.model.lgb_random_state,
-                    n_jobs=-1,
-                    verbose=-1,
-                )
-        # モデル訓練
-        self.model.fit(X_train_scaled, y_train)
-        # 予測と評価
-        train_pred = self.model.predict(X_train_scaled)
-        test_pred = self.model.predict(X_test_scaled)
-        if task_type == "regression":
-            train_mse = mean_squared_error(y_train, train_pred)
-            test_mse = mean_squared_error(y_test, test_pred)
-            logger.info(f"Training MSE: {train_mse:.4f}")
-            logger.info(f"Test MSE: {test_mse:.4f}")
-        else:
-            train_acc = accuracy_score(y_train, train_pred)
-            test_acc = accuracy_score(y_test, test_pred)
-            logger.info(f"Training Accuracy: {train_acc:.4f}")
-            logger.info(f"Test Accuracy: {test_acc:.4f}")
-        self.is_trained = True
-        logger.info("Model training completed")
-        # モデルを保存
-        self.save_model()
-
-    def predict_score(self, symbol: str) -> float:
-        """単一銘柄のスコアを予測"""
-        if not self.is_trained:
-            logger.warning("Model not trained, loading from file...")
-            if not self.load_model():
-                logger.error("No trained model available")
-                return 50.0
-        try:
-            # データ取得
-            data = self.data_provider.get_stock_data(symbol, "1y")
-            if data.empty:
-                return 0.0
-            # 特徴量準備
-            features = self.prepare_features(data)
-            if features.empty:
-                return 0.0
-            # 最新データの特徴量を取得
-            latest_features = features.iloc[-1:].copy()
-            # 銘柄ワンホットエンコーディング
-            for feature_name in self.feature_names:
-                if feature_name.startswith("symbol_"):
-                    latest_features[feature_name] = (
-                        1 if feature_name == f"symbol_{symbol}" else 0
-                    )
-                elif feature_name not in latest_features.columns:
-                    latest_features[feature_name] = 0
-            # 特徴量順序を合わせる
-            latest_features = latest_features.reindex(
-                columns=self.feature_names, fill_value=0
-            )
-            # スケーリング
-            features_scaled = self.scaler.transform(latest_features)
-            # 予測
-            score = self.model.predict(features_scaled)[0]
-            # スコアを0-100に正規化
-            score = max(0, min(100, float(score)))
-            return score
-        except Exception as e:
-            logger.error(f"Error predicting score for {symbol}: {str(e)}")
-            return 50.0
-
-    def predict_return_rate(self, symbol: str, days: int = 5) -> float:
-        """リターン率を直接予測（改善されたMAPE対応）"""
-        if not self.is_trained:
-            logger.warning("Model not trained for return prediction")
-            return 0.0
-        try:
-            # データ取得
-            data = self.data_provider.get_stock_data(symbol, "1y")
-            if data.empty or len(data) < 50:
-                return 0.0
-            # 特徴量準備
-            features = self.prepare_features_for_return_prediction(data)
-            if features.empty:
-                return 0.0
-            # 最新データの特徴量を取得
-            latest_features = features.iloc[-1:].copy()
-            # 銘柄ワンホットエンコーディング
-            for feature_name in self.feature_names:
-                if feature_name.startswith("symbol_"):
-                    latest_features[feature_name] = (
-                        1 if feature_name == f"symbol_{symbol}" else 0
-                    )
-                elif feature_name not in latest_features.columns:
-                    latest_features[feature_name] = 0
-            # 特徴量順序を合わせる
-            latest_features = latest_features.reindex(
-                columns=self.feature_names, fill_value=0
-            )
-            # スケーリング
-            features_scaled = self.scaler.transform(latest_features)
-            # 予測（リターン率）
-            predicted_return = self.model.predict(features_scaled)[0]
-            # 現実的な範囲に制限（日数に応じて調整）
-            max_return = 0.006 * days  # 1日あたり最大0.6%
-            predicted_return = max(
-                -max_return, min(max_return, float(predicted_return))
-            )
-            return predicted_return
-        except Exception as e:
-            logger.error(f"Error predicting return rate for {symbol}: {str(e)}")
-            return 0.0
-
-    def prepare_features_for_return_prediction(
-        self, data: pd.DataFrame
-    ) -> pd.DataFrame:
-        """リターン率予測用の特徴量準備"""
-        features = self.prepare_features(data)
-        # リターン率関連の特徴量を追加
-        if len(data) >= 10:
-            returns = data["Close"].pct_change()
-            # 過去のリターン率統計
-            features["return_mean_5d"] = returns.rolling(5).mean()
-            features["return_std_5d"] = returns.rolling(5).std()
-            features["return_mean_20d"] = returns.rolling(20).mean()
-            features["return_std_20d"] = returns.rolling(20).std()
-            # シャープレシオ風指標
-            features["return_sharpe_5d"] = features["return_mean_5d"] / (
-                features["return_std_5d"] + 1e-8
-            )
-            features["return_sharpe_20d"] = features["return_mean_20d"] / (
-                features["return_std_20d"] + 1e-8
-            )
-            # 連続上昇/下降日数
-            returns_sign = np.sign(returns)
-            features["consecutive_up"] = (
-                returns_sign.groupby(
-                    (returns_sign != returns_sign.shift()).cumsum()
-                ).cumcount()
-                + 1
-            ) * (returns_sign > 0)
-            features["consecutive_down"] = (
-                returns_sign.groupby(
-                    (returns_sign != returns_sign.shift()).cumsum()
-                ).cumcount()
-                + 1
-            ) * (returns_sign < 0)
-        return features.fillna(0)
-
-    def get_feature_importance(self) -> Dict[str, float]:
-        """特徴量重要度を取得"""
-        if not self.is_trained or self.model is None:
-            return {}
-        try:
-            if hasattr(self.model, "feature_importances_"):
-                importances = self.model.feature_importances_
-                return dict(zip(self.feature_names, importances))
-        except Exception as e:
-            logger.error(f"Error getting feature importance: {str(e)}")
-        return {}
-
-    def save_model(self):
-        """モデルを保存"""
-        try:
-            model_file = (
-                self.model_path / f"ml_stock_predictor_{self.model_type}.joblib"
-            )
-            scaler_file = self.model_path / f"scaler_{self.model_type}.joblib"
-            features_file = self.model_path / f"features_{self.model_type}.joblib"
-            joblib.dump(self.model, model_file)
-            joblib.dump(self.scaler, scaler_file)
-            joblib.dump(self.feature_names, features_file)
-            logger.info(f"Model saved to {model_file}")
-        except Exception as e:
-            logger.error(f"Error saving model: {str(e)}")
-
-    def load_model(self) -> bool:
-        """モデルを読み込み"""
-        try:
-            model_file = (
-                self.model_path / f"ml_stock_predictor_{self.model_type}.joblib"
-            )
-            scaler_file = self.model_path / f"scaler_{self.model_type}.joblib"
-            features_file = self.model_path / f"features_{self.model_type}.joblib"
-            if not all(
-                [model_file.exists(), scaler_file.exists(), features_file.exists()]
-            ):
-                return False
-            self.model = joblib.load(model_file)
-            self.scaler = joblib.load(scaler_file)
-            self.feature_names = joblib.load(features_file)
-            self.is_trained = True
-            logger.info(f"Model loaded from {model_file}")
-            return True
-        except Exception as e:
-            logger.error(f"Error loading model: {str(e)}")
-            return False
-
-
-class EnsembleStockPredictor:
-    """複数モデルのアンサンブル予測器"""
-
-    def __init__(self):
-        self.models = {}
-        self.weights = {}
-        self.data_provider = StockDataProvider()
-        self.is_trained = False
-        self.feature_names = []
-        self.model_path = Path("models/saved_models")
-        self.model_path.mkdir(exist_ok=True)
-
-    def add_model(self, name: str, model, weight: float = 1.0):
-        """アンサンブルにモデルを追加"""
-        self.models[name] = model
-        self.weights[name] = weight
-
-    def prepare_ensemble_models(self):
-        """複数のモデルタイプを準備"""
-        from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
-        from sklearn.neural_network import MLPRegressor
-        
-
-        # XGBoost
-        xgb_model = xgb.XGBRegressor(
-            n_estimators=200,
-            max_depth=6,
-            learning_rate=0.1,
-            subsample=0.8,
-            colsample_bytree=0.8,
-            random_state=42,
-            n_jobs=-1,
-        )
-        # LightGBM
-        lgb_model = lgb.LGBMRegressor(
-            n_estimators=200,
-            max_depth=6,
-            learning_rate=0.1,
-            subsample=0.8,
-            colsample_bytree=0.8,
-            random_state=42,
-            n_jobs=-1,
-            verbose=-1,
-        )
-        # Random Forest
-        rf_model = RandomForestRegressor(
-            n_estimators=100, max_depth=10, random_state=42, n_jobs=-1
-        )
-        # Gradient Boosting
-        gb_model = GradientBoostingRegressor(
-            n_estimators=100, max_depth=6, learning_rate=0.1, random_state=42
-        )
-        # Neural Network
-        nn_model = MLPRegressor(
-            hidden_layer_sizes=(100, 50), max_iter=500, random_state=42
-        )
-        # アンサンブルに追加（重み付け）
-        self.add_model("xgboost", xgb_model, weight=0.3)
-        self.add_model("lightgbm", lgb_model, weight=0.3)
-        self.add_model("random_forest", rf_model, weight=0.2)
-        self.add_model("gradient_boost", gb_model, weight=0.15)
-        self.add_model("neural_network", nn_model, weight=0.05)
-
-    def train_ensemble(
-        self, symbols: List[str], target_column: str = "recommendation_score"
-    ):
-        """アンサンブルモデルを訓練"""
-        from config.settings import get_settings
-
-        settings = get_settings()
-        # モデル準備
-        self.prepare_ensemble_models()
-        # 単一モデルインスタンスでデータ準備
-        ml_predictor = MLStockPredictor()
-        logger.info("Preparing dataset for ensemble...")
-        features, targets_reg, targets_cls = ml_predictor.prepare_dataset(symbols)
-        if target_column not in targets_reg.columns:
-            raise ValueError(f"Target column {target_column} not found")
-        targets = targets_reg[target_column]
-        # 欠損値除去
-        valid_indices = ~(targets.isna() | features.isna().any(axis=1))
-        features_clean = features[valid_indices]
-        targets_clean = targets[valid_indices]
-        if len(features_clean) < settings.model.min_training_data:
-            raise ValueError(f"Insufficient training data: {len(features_clean)}")
-        self.feature_names = features_clean.columns.tolist()
-        # 時系列分割
-        train_size = int(len(features_clean) * settings.model.train_test_split)
-        X_train = features_clean.iloc[:train_size]
-        X_test = features_clean.iloc[train_size:]
-        y_train = targets_clean.iloc[:train_size]
-        y_test = targets_clean.iloc[train_size:]
-        # 特徴量スケーリング
-        scaler = StandardScaler()
-        X_train_scaled = scaler.fit_transform(X_train)
-        X_test_scaled = scaler.transform(X_test)
-        # 各モデルを訓練
-        model_predictions = {}
-        model_scores = {}
-        for name, model in self.models.items():
-            logger.info(f"Training {name}...")
-            try:
-                model.fit(X_train_scaled, y_train)
-                # 予測と評価
-                train_pred = model.predict(X_train_scaled)
-                test_pred = model.predict(X_test_scaled)
-                train_mse = mean_squared_error(y_train, train_pred)
-                test_mse = mean_squared_error(y_test, test_pred)
-                model_predictions[name] = test_pred
-                model_scores[name] = test_mse
-                logger.info(
-                    f"{name} - Train MSE: {train_mse:.4f}, Test MSE: {test_mse:.4f}"
-                )
-            except Exception as e:
-                logger.error(f"Error training {name}: {str(e)}")
-                # 失敗したモデルは除外
-                del self.models[name]
-                del self.weights[name]
-        # 動的重み調整（性能に基づく）
-        self._adjust_weights_based_on_performance(model_scores)
-        # アンサンブル予測の評価
-        ensemble_pred = self._ensemble_predict_from_predictions(model_predictions)
-        ensemble_mse = mean_squared_error(y_test, ensemble_pred)
-        logger.info(f"Ensemble MSE: {ensemble_mse:.4f}")
-        logger.info(f"Final model weights: {self.weights}")
-        self.scaler = scaler
-        self.is_trained = True
-        # モデル保存
-        self.save_ensemble()
-
-    def _adjust_weights_based_on_performance(self, model_scores: Dict[str, float]):
-        """性能に基づいて重みを動的調整"""
-        # MSEが低いほど良いので、逆数を取って重み計算
-        inverse_scores = {
-            name: 1.0 / (score + 1e-6) for name, score in model_scores.items()
-        }
-        total_inverse = sum(inverse_scores.values())
-        # 正規化して新しい重みを設定
-        for name in self.weights:
-            if name in inverse_scores:
-                self.weights[name] = inverse_scores[name] / total_inverse
-
-    def _ensemble_predict_from_predictions(
-        self, model_predictions: Dict[str, np.ndarray]
-    ) -> np.ndarray:
-        """複数モデルの予測を重み付き平均"""
-        weighted_sum = np.zeros_like(list(model_predictions.values())[0])
-        total_weight = 0
-        for name, predictions in model_predictions.items():
-            if name in self.weights:
-                weighted_sum += predictions * self.weights[name]
-                total_weight += self.weights[name]
-        return weighted_sum / total_weight if total_weight > 0 else weighted_sum
-
-    def predict_score(self, symbol: str) -> float:
-        """アンサンブル予測"""
-        if not self.is_trained:
-            if not self.load_ensemble():
-                logger.error("No trained ensemble model available")
-                return 50.0
-        try:
-            # データ取得と特徴量準備
-            data = self.data_provider.get_stock_data(symbol, "1y")
-            if data.empty:
-                return 0.0
-            ml_predictor = MLStockPredictor()
-            features = ml_predictor.prepare_features(data)
-            if features.empty:
-                return 0.0
-            # 最新データの特徴量
-            latest_features = features.iloc[-1:].copy()
-            # 特徴量を訓練時と同じ順序に調整
-            latest_features = latest_features.reindex(
-                columns=self.feature_names, fill_value=0
-            )
-            # スケーリング
-            features_scaled = self.scaler.transform(latest_features)
-            # 各モデルの予測を収集
-            predictions = {}
-            for name, model in self.models.items():
-                try:
-                    pred = model.predict(features_scaled)[0]
-                    predictions[name] = pred
-                except Exception as e:
-                    logger.warning(f"Error with {name} prediction: {str(e)}")
-            # アンサンブル予測
-            if predictions:
-                ensemble_score = self._ensemble_predict_from_predictions(
-                    {name: np.array([pred]) for name, pred in predictions.items()}
-                )[0]
-                return max(0, min(100, float(ensemble_score)))
-            else:
-                return 50.0
-        except Exception as e:
-            logger.error(f"Error in ensemble prediction for {symbol}: {str(e)}")
-            return 50.0
-
-    def save_ensemble(self):
-        """アンサンブルモデルを保存"""
-        try:
-            ensemble_file = self.model_path / "ensemble_models.joblib"
-            ensemble_data = {
-                "models": self.models,
-                "weights": self.weights,
-                "scaler": self.scaler,
-                "feature_names": self.feature_names,
-                "is_trained": self.is_trained,
-            }
-            joblib.dump(ensemble_data, ensemble_file)
-            logger.info(f"Ensemble saved to {ensemble_file}")
-        except Exception as e:
-            logger.error(f"Error saving ensemble: {str(e)}")
-
-    def load_ensemble(self) -> bool:
-        """アンサンブルモデルを読み込み"""
-        try:
-            ensemble_file = self.model_path / "ensemble_models.joblib"
-            if not ensemble_file.exists():
-                return False
-            ensemble_data = joblib.load(ensemble_file)
-            self.models = ensemble_data["models"]
-            self.weights = ensemble_data["weights"]
-            self.scaler = ensemble_data["scaler"]
-            self.feature_names = ensemble_data["feature_names"]
-            self.is_trained = ensemble_data["is_trained"]
-            logger.info(f"Ensemble loaded from {ensemble_file}")
-            return True
-        except Exception as e:
-            logger.error(f"Error loading ensemble: {str(e)}")
-            return False
-
-
-class HyperparameterOptimizer:
-    """ハイパーパラメータ自動調整"""
-
-    def __init__(self):
-        self.best_params = {}
-        self.optimization_history = []
-
-    def optimize_xgboost(self, X, y, cv_folds=5, n_trials=100):
-        """XGBoostパラメータ最適化"""
-        import optuna
-
-        def objective(trial):
-            params = {
-                "n_estimators": trial.suggest_int("n_estimators", 100, 500),
-                "max_depth": trial.suggest_int("max_depth", 3, 10),
-                "learning_rate": trial.suggest_float("learning_rate", 0.01, 0.3),
-                "subsample": trial.suggest_float("subsample", 0.6, 1.0),
-                "colsample_bytree": trial.suggest_float("colsample_bytree", 0.6, 1.0),
-                "gamma": trial.suggest_float("gamma", 0, 5),
-                "min_child_weight": trial.suggest_int("min_child_weight", 1, 10),
-                "random_state": 42,
-                "n_jobs": -1,
-            }
-            model = xgb.XGBRegressor(**params)
-            scores = cross_val_score(
-                model, X, y, cv=cv_folds, scoring="neg_mean_squared_error"
-            )
-            return -scores.mean()
-
-        study = optuna.create_study(direction="minimize")
-        study.optimize(objective, n_trials=n_trials)
-        self.best_params["xgboost"] = study.best_params
-        logger.info(f"Best XGBoost params: {study.best_params}")
-        logger.info(f"Best XGBoost score: {study.best_value}")
-        return study.best_params
-
-    def optimize_lightgbm(self, X, y, cv_folds=5, n_trials=100):
-        """LightGBMパラメータ最適化"""
-        import optuna
-
-        def objective(trial):
-            params = {
-                "n_estimators": trial.suggest_int("n_estimators", 100, 500),
-                "max_depth": trial.suggest_int("max_depth", 3, 10),
-                "learning_rate": trial.suggest_float("learning_rate", 0.01, 0.3),
-                "subsample": trial.suggest_float("subsample", 0.6, 1.0),
-                "colsample_bytree": trial.suggest_float("colsample_bytree", 0.6, 1.0),
-                "min_child_samples": trial.suggest_int("min_child_samples", 5, 100),
-                "reg_alpha": trial.suggest_float("reg_alpha", 0, 1),
-                "reg_lambda": trial.suggest_float("reg_lambda", 0, 1),
-                "random_state": 42,
-                "n_jobs": -1,
-                "verbose": -1,
-            }
-            model = lgb.LGBMRegressor(**params)
-            scores = cross_val_score(
-                model, X, y, cv=cv_folds, scoring="neg_mean_squared_error"
-            )
-            return -scores.mean()
-
-        study = optuna.create_study(direction="minimize")
-        study.optimize(objective, n_trials=n_trials)
-        self.best_params["lightgbm"] = study.best_params
-        logger.info(f"Best LightGBM params: {study.best_params}")
-        logger.info(f"Best LightGBM score: {study.best_value}")
-        return study.best_params
-
-    def save_best_params(self):
-        """最適パラメータを保存"""
-        try:
-            params_file = Path("models/saved_models/best_hyperparams.json")
-            with open(params_file, "w") as f:
-                import json
-
-                json.dump(self.best_params, f, indent=2)
-            logger.info(f"Best parameters saved to {params_file}")
-        except Exception as e:
-            logger.error(f"Error saving hyperparameters: {str(e)}")
-
-    def load_best_params(self):
-        """最適パラメータを読み込み"""
-        try:
-            params_file = Path("models/saved_models/best_hyperparams.json")
-            if params_file.exists():
-                with open(params_file, "r") as f:
-                    import json
-
-                    self.best_params = json.load(f)
-                logger.info("Best parameters loaded")
-                return True
-            return False
-        except Exception as e:
-            logger.error(f"Error loading hyperparameters: {str(e)}")
-            return False
-
-
-class ModelPerformanceMonitor:
-    """モデル性能監視・評価システム"""
-
-    def __init__(self):
-        self.performance_history = []
-        self.alerts = []
-
-    def evaluate_model_performance(self, model, X_test, y_test, model_name="Unknown"):
-        """詳細なモデル性能評価"""
-        predictions = model.predict(X_test)
-        # 基本メトリクス
-        mse = mean_squared_error(y_test, predictions)
-        rmse = np.sqrt(mse)
-        mae = mean_absolute_error(y_test, predictions)
-        r2 = r2_score(y_test, predictions)
-        # 方向精度（上昇/下降予測の正確性）
-        y_direction = (y_test > 50).astype(int)
-        pred_direction = (predictions > 50).astype(int)
-        direction_accuracy = accuracy_score(y_direction, pred_direction)
-        # 分位点ごとの性能
-        quantiles = np.quantile(y_test, [0.25, 0.5, 0.75])
-        quantile_performance = {}
-        for i, q in enumerate([0.25, 0.5, 0.75]):
-            mask = (
-                y_test <= quantiles[i]
-                if i == 0
-                else (y_test > quantiles[i - 1]) & (y_test <= quantiles[i])
-            )
-            if mask.sum() > 0:
-                quantile_mse = mean_squared_error(y_test[mask], predictions[mask])
-                quantile_performance[f"Q{i+1}"] = quantile_mse
-        # 性能記録
-        performance_record = {
-            "timestamp": datetime.now().isoformat(),
-            "model_name": model_name,
-            "mse": mse,
-            "rmse": rmse,
-            "mae": mae,
-            "r2_score": r2,
-            "direction_accuracy": direction_accuracy,
-            "quantile_performance": quantile_performance,
-            "sample_size": len(y_test),
-        }
-        self.performance_history.append(performance_record)
-        # アラート判定
-        self._check_performance_alerts(performance_record)
-        logger.info(f"Model {model_name} Performance:")
-        logger.info(f"  RMSE: {rmse:.4f}")
-        logger.info(f"  MAE: {mae:.4f}")
-        logger.info(f"  R²: {r2:.4f}")
-        logger.info(f"  Direction Accuracy: {direction_accuracy:.4f}")
-        return performance_record
-
-    def _check_performance_alerts(self, performance_record):
-        """性能低下アラートをチェック"""
-        rmse_threshold = 15.0
-        r2_threshold = 0.1
-        direction_threshold = 0.55
-        alerts = []
-        if performance_record["rmse"] > rmse_threshold:
-            alerts.append(
-                f"High RMSE: {performance_record['rmse']:.4f} > {rmse_threshold}"
-            )
-        if performance_record["r2_score"] < r2_threshold:
-            alerts.append(
-                f"Low R²: {performance_record['r2_score']:.4f} < {r2_threshold}"
-            )
-        if performance_record["direction_accuracy"] < direction_threshold:
-            alerts.append(
-                f"Low Direction Accuracy: {performance_record['direction_accuracy']:.4f} < {direction_threshold}"
-            )
-        if alerts:
-            alert_record = {
-                "timestamp": performance_record["timestamp"],
-                "model_name": performance_record["model_name"],
-                "alerts": alerts,
-            }
-            self.alerts.append(alert_record)
-            logger.warning(
-                f"Performance alerts for {performance_record['model_name']}: {alerts}"
-            )
-
-    def get_performance_summary(self, last_n_records=10):
-        """性能サマリーを取得"""
-        if not self.performance_history:
-            return "No performance data available"
-        recent_records = self.performance_history[-last_n_records:]
-        avg_rmse = np.mean([r["rmse"] for r in recent_records])
-        avg_r2 = np.mean([r["r2_score"] for r in recent_records])
-        avg_direction = np.mean([r["direction_accuracy"] for r in recent_records])
-        summary = f"""
-Performance Summary (Last {len(recent_records)} records):
-  Average RMSE: {avg_rmse:.4f}
-  Average R²: {avg_r2:.4f}
-  Average Direction Accuracy: {avg_direction:.4f}
-  Total Alerts: {len(self.alerts)}
-        """
-        return summary
-
-    def save_performance_data(self):
-        """性能データを保存"""
-        try:
-            perf_file = Path("models/saved_models/performance_history.json")
-            data = {
-                "performance_history": self.performance_history,
-                "alerts": self.alerts,
-            }
-            with open(perf_file, "w") as f:
-                import json
-
-                json.dump(data, f, indent=2)
-            logger.info(f"Performance data saved to {perf_file}")
-        except Exception as e:
-            logger.error(f"Error saving performance data: {str(e)}")
-
-    def load_performance_data(self):
-        """性能データを読み込み"""
-        try:
-            perf_file = Path("models/saved_models/performance_history.json")
-            if perf_file.exists():
-                with open(perf_file, "r") as f:
-                    import json
-
-                    data = json.load(f)
-                    self.performance_history = data.get("performance_history", [])
-                    self.alerts = data.get("alerts", [])
-                logger.info("Performance data loaded")
-                return True
-            return False
-        except Exception as e:
-            logger.error(f"Error loading performance data: {str(e)}")
-            return False
-
-
-class ParallelStockPredictor:
-    """並列処理対応の高速株価予測器"""
-
-    def __init__(self, ensemble_predictor: EnsembleStockPredictor, n_jobs: int = -1):
-        self.ensemble_predictor = ensemble_predictor
-        self.n_jobs = n_jobs if n_jobs != -1 else os.cpu_count()
-        self.batch_cache = {}
-
-    def predict_multiple_stocks_parallel(self, symbols: List[str]) -> Dict[str, float]:
-        """複数銘柄の並列予測"""
-        from concurrent.futures import ThreadPoolExecutor, as_completed
-
-        results = {}
-        # キャッシュチェック
-        uncached_symbols = []
-        for symbol in symbols:
-            if symbol in self.batch_cache:
-                results[symbol] = self.batch_cache[symbol]
-            else:
-                uncached_symbols.append(symbol)
-        if not uncached_symbols:
-            return results
-        logger.info(f"Predicting {len(uncached_symbols)} stocks in parallel...")
-        # 並列実行
-        with ThreadPoolExecutor(max_workers=self.n_jobs) as executor:
-            future_to_symbol = {
-                executor.submit(self.ensemble_predictor.predict_score, symbol): symbol
-                for symbol in uncached_symbols
-            }
-            for future in as_completed(future_to_symbol):
-                symbol = future_to_symbol[future]
-                try:
-                    score = future.result()
-                    results[symbol] = score
-                    self.batch_cache[symbol] = score
-                except Exception as e:
-                    logger.error(f"Error predicting {symbol}: {str(e)}")
-                    results[symbol] = 50.0
-        return results
-
-    def batch_data_preparation(self, symbols: List[str]) -> Dict[str, pd.DataFrame]:
-        """バッチデータ準備（並列）"""
-        from concurrent.futures import ThreadPoolExecutor, as_completed
-
-        data_results = {}
-        with ThreadPoolExecutor(max_workers=self.n_jobs) as executor:
-            future_to_symbol = {
-                executor.submit(self._get_stock_data_safe, symbol): symbol
-                for symbol in symbols
-            }
-            for future in as_completed(future_to_symbol):
-                symbol = future_to_symbol[future]
-                try:
-                    data = future.result()
-                    if not data.empty:
-                        data_results[symbol] = data
-                except Exception as e:
-                    logger.error(f"Error getting data for {symbol}: {str(e)}")
-        return data_results
-
-    def _get_stock_data_safe(self, symbol: str) -> pd.DataFrame:
-        """安全なデータ取得"""
-        try:
-            return self.ensemble_predictor.data_provider.get_stock_data(symbol, "1y")
-        except Exception as e:
-            logger.error(f"Error fetching data for {symbol}: {str(e)}")
-            return pd.DataFrame()
-
-    def clear_batch_cache(self):
-        """バッチキャッシュをクリア"""
-        self.batch_cache.clear()
-        logger.info("Batch cache cleared")
-
-
-class AdvancedCacheManager:
-    """高度なキャッシュ管理システム"""
-
-    def __init__(self):
-        self.feature_cache = {}
-        self.prediction_cache = {}
-        self.cache_stats = {
-            "hits": 0,
-            "misses": 0,
-            "feature_cache_size": 0,
-            "prediction_cache_size": 0,
-        }
-
-    def get_cached_features(
-        self, symbol: str, data_hash: str
-    ) -> Optional[pd.DataFrame]:
-        """特徴量キャッシュから取得"""
-        cache_key = f"{symbol}_{data_hash}"
-        if cache_key in self.feature_cache:
-            self.cache_stats["hits"] += 1
-            return self.feature_cache[cache_key]
-        else:
-            self.cache_stats["misses"] += 1
-            return None
-
-    def cache_features(self, symbol: str, data_hash: str, features: pd.DataFrame):
-        """特徴量をキャッシュ"""
-        cache_key = f"{symbol}_{data_hash}"
-        self.feature_cache[cache_key] = features
-        self.cache_stats["feature_cache_size"] = len(self.feature_cache)
-
-    def get_cached_prediction(self, symbol: str, features_hash: str) -> Optional[float]:
-        """予測結果キャッシュから取得"""
-        cache_key = f"{symbol}_{features_hash}"
-        if cache_key in self.prediction_cache:
-            self.cache_stats["hits"] += 1
-            return self.prediction_cache[cache_key]
-        else:
-            self.cache_stats["misses"] += 1
-            return None
-
-    def cache_prediction(self, symbol: str, features_hash: str, prediction: float):
-        """予測結果をキャッシュ"""
-        cache_key = f"{symbol}_{features_hash}"
-        self.prediction_cache[cache_key] = prediction
-        self.cache_stats["prediction_cache_size"] = len(self.prediction_cache)
-
-    def cleanup_old_cache(self, max_size: int = 1000):
-        """古いキャッシュをクリーンアップ"""
-        if len(self.feature_cache) > max_size:
-            # 最も古いエントリを削除（簡単なLRU実装）
-            keys_to_remove = list(self.feature_cache.keys())[:-max_size]
-            for key in keys_to_remove:
-                del self.feature_cache[key]
-        if len(self.prediction_cache) > max_size:
-            keys_to_remove = list(self.prediction_cache.keys())[:-max_size]
-            for key in keys_to_remove:
-                del self.prediction_cache[key]
-        self.cache_stats["feature_cache_size"] = len(self.feature_cache)
-        self.cache_stats["prediction_cache_size"] = len(self.prediction_cache)
-
-    def get_cache_stats(self) -> Dict:
-        """キャッシュ統計を取得"""
-        total_requests = self.cache_stats["hits"] + self.cache_stats["misses"]
-        hit_rate = (
-            self.cache_stats["hits"] / total_requests if total_requests > 0 else 0
-        )
-        return {
-            **self.cache_stats,
-            "hit_rate": hit_rate,
-            "total_requests": total_requests,
-        }
-
-
-class DeepLearningPredictor:
-    """LSTM/Transformer深層学習予測器"""
-
-    def __init__(self, model_type="lstm"):
-        self.model_type = model_type
-        self.model = None
-        self.scaler = StandardScaler()
-        self.sequence_length = 60
-        self.is_trained = False
-        self.feature_columns = []
-
-    def prepare_sequences(
-        self, data: pd.DataFrame, target_col: str = "Close"
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        """時系列データをシーケンスに変換"""
-        # 特徴量とターゲット分離
-        feature_data = data.drop(["Close"], axis=1).values
-        target_data = data[target_col].values
-        # 正規化
-        feature_data = self.scaler.fit_transform(feature_data)
-        X, y = [], []
-        for i in range(self.sequence_length, len(feature_data)):
-            X.append(feature_data[i - self.sequence_length : i])
-            y.append(target_data[i])
-        return np.array(X), np.array(y)
-
-    def build_lstm_model(self, input_shape):
-        """LSTM モデル構築"""
-        import tensorflow as tf
-        from tensorflow.keras.models import Sequential
-        from tensorflow.keras.layers import LSTM, Dense, Dropout, BatchNormalization
-        from tensorflow.keras.optimizers import Adam
-
-        model = Sequential(
-            [
-                LSTM(128, return_sequences=True, input_shape=input_shape),
-                Dropout(0.3),
-                BatchNormalization(),
-                LSTM(128, return_sequences=True),
-                Dropout(0.3),
-                BatchNormalization(),
-                LSTM(64, return_sequences=False),
-                Dropout(0.3),
-                BatchNormalization(),
-                Dense(32, activation="relu"),
-                Dropout(0.2),
-                Dense(1, activation="linear"),
-            ]
-        )
-        model.compile(optimizer=Adam(learning_rate=0.001), loss="mse", metrics=["mae"])
-        return model
-
-    def build_transformer_model(self, input_shape):
-        """Transformer モデル構築"""
-        import tensorflow as tf
-        from tensorflow.keras import layers
-
-        def transformer_encoder(inputs, head_size, num_heads, ff_dim, dropout=0):
-            # Multi-head self-attention
-            x = layers.MultiHeadAttention(
-                key_dim=head_size, num_heads=num_heads, dropout=dropout
-            )(inputs, inputs)
-            x = layers.Dropout(dropout)(x)
-            x = layers.LayerNormalization(epsilon=1e-6)(x)
-            res = x + inputs
-            # Feed forward network
-            x = layers.Conv1D(filters=ff_dim, kernel_size=1, activation="relu")(res)
-            x = layers.Dropout(dropout)(x)
-            x = layers.Conv1D(filters=inputs.shape[-1], kernel_size=1)(x)
-            x = layers.LayerNormalization(epsilon=1e-6)(x)
-            return x + res
-
-        inputs = tf.keras.Input(shape=input_shape)
-        x = inputs
-        # Multi-layer transformer
-        for _ in range(3):
-            x = transformer_encoder(
-                x, head_size=64, num_heads=4, ff_dim=128, dropout=0.3
-            )
-        x = layers.GlobalAveragePooling1D(data_format="channels_first")(x)
-        x = layers.Dropout(0.3)(x)
-        x = layers.Dense(32, activation="relu")(x)
-        x = layers.Dropout(0.2)(x)
-        outputs = layers.Dense(1, activation="linear")(x)
-        model = tf.keras.Model(inputs, outputs)
-        model.compile(optimizer="adam", loss="mse", metrics=["mae"])
-        return model
-
-    def train_deep_model(self, symbols: List[str]):
-        """深層学習モデル訓練"""
-        # データ準備
-        all_data = []
-        ml_predictor = MLStockPredictor()
-        for symbol in symbols:
-            data = ml_predictor.data_provider.get_stock_data(symbol, "3y")
-            if len(data) < 200:
-                continue
-            features = ml_predictor.prepare_features(data)
-            # Closeカラムを追加
-            features["Close"] = data["Close"]
-            all_data.append(features.dropna())
-        if not all_data:
-            raise ValueError("No sufficient data for training")
-        # 全データ結合
-        combined_data = pd.concat(all_data, ignore_index=True)
-        # シーケンス作成
-        X, y = self.prepare_sequences(combined_data)
-        # 訓練/テスト分割
-        split_idx = int(len(X) * 0.8)
-        X_train, X_test = X[:split_idx], X[split_idx:]
-        y_train, y_test = y[:split_idx], y[split_idx:]
-        # モデル構築
-        input_shape = (X_train.shape[1], X_train.shape[2])
-        if self.model_type == "lstm":
-            self.model = self.build_lstm_model(input_shape)
-        else:  # transformer
-            self.model = self.build_transformer_model(input_shape)
-        # 訓練
-        from tensorflow.keras.callbacks import EarlyStopping, ReduceLROnPlateau
-
-        callbacks = [
-            EarlyStopping(patience=15, restore_best_weights=True),
-            ReduceLROnPlateau(patience=7, factor=0.5),
-        ]
-        history = self.model.fit(
-            X_train,
-            y_train,
-            validation_data=(X_test, y_test),
-            epochs=100,
-            batch_size=32,
-            callbacks=callbacks,
-            verbose=1,
-        )
-        self.is_trained = True
-        self.save_deep_model()
-        # 評価
-        test_pred = self.model.predict(X_test)
-        test_mse = mean_squared_error(y_test, test_pred)
-        logger.info(f"Deep Learning {self.model_type} Test MSE: {test_mse:.4f}")
-        return history
-
-    def predict_deep(self, symbol: str) -> float:
-        """深層学習予測"""
-        if not self.is_trained:
-            return 50.0
-        try:
-            ml_predictor = MLStockPredictor()
-            data = ml_predictor.data_provider.get_stock_data(symbol, "1y")
-            features = ml_predictor.prepare_features(data)
-            features["Close"] = data["Close"]
-            if len(features) < self.sequence_length:
-                return 50.0
-            # 最新シーケンス準備
-            recent_data = (
-                features.tail(self.sequence_length).drop(["Close"], axis=1).values
-            )
-            recent_data = self.scaler.transform(recent_data)
-            sequence = recent_data.reshape(1, self.sequence_length, -1)
-            # 予測
-            pred = self.model.predict(sequence)[0][0]
-            # スコア変換 (価格予測→0-100スコア)
-            current_price = data["Close"].iloc[-1]
-            score = 50 + (pred - current_price) / current_price * 100
-            return max(0, min(100, score))
-        except Exception as e:
-            logger.error(f"Deep learning prediction error for {symbol}: {str(e)}")
-            return 50.0
-
-    def save_deep_model(self):
-        """深層学習モデル保存"""
-        try:
-            model_path = Path("models/saved_models")
-            self.model.save(model_path / f"deep_{self.model_type}_model.h5")
-            joblib.dump(
-                self.scaler, model_path / f"deep_{self.model_type}_scaler.joblib"
-            )
-            logger.info(f"Deep {self.model_type} model saved")
-        except Exception as e:
-            logger.error(f"Error saving deep model: {str(e)}")
-
-
-class AdvancedEnsemblePredictor:
-    """
-    84.6%精度突破を目指す高度アンサンブル学習システム
-    特徴:
-    - BERT活用ニュースセンチメント分析
-    - マクロ経済指標統合
-    - 時系列Transformer最適化
-    - 動的重み調整アンサンブル
-    """
-
-    def __init__(self):
-        self.models = {}
-        self.weights = {}
-        self.macro_data = {}
-        self.sentiment_analyzer = None
-        self.confidence_threshold = 0.75
-        self.logger = logging.getLogger(__name__)
-
-    def initialize_components(self):
-        """各コンポーネント初期化"""
-        try:
-            # 基本予測モデル群
-            self.models = {
-                "trend_following": None,  # 84.6%ベースモデル
-                "lstm_deep": DeepLearningPredictor("lstm"),
-                "transformer_deep": DeepLearningPredictor("transformer"),
-                "sentiment_enhanced": None,
-                "macro_enhanced": None,
-            }
-            # 初期重み設定（84.6%モデルを重視）
-            self.weights = {
-                "trend_following": 0.4,  # 84.6%の実績重視
-                "lstm_deep": 0.2,
-                "transformer_deep": 0.2,
-                "sentiment_enhanced": 0.1,
-                "macro_enhanced": 0.1,
-            }
-            # センチメント分析器初期化
-            self._initialize_sentiment_analyzer()
-            # マクロ経済データ取得
-            self._initialize_macro_data()
-        except Exception as e:
-            self.logger.error(f"コンポーネント初期化エラー: {e}")
-
-    def _initialize_sentiment_analyzer(self):
-        """BERT活用センチメント分析器初期化"""
-        try:
-            # transformersライブラリが利用可能な場合のみ
-            from transformers import BertTokenizer, BertForSequenceClassification
-
-            # 日本語BERT事前学習モデル
-            model_name = "cl-tohoku/bert-base-japanese-whole-word-masking"
-            # セキュリティ向上: 特定のリビジョンを指定
-            revision = "f012345678901234567890123456789012345678"  # 特定のコミットハッシュ
-            self.tokenizer = BertTokenizer.from_pretrained(model_name, revision=revision)  # nosec B615
-            self.bert_model = BertForSequenceClassification.from_pretrained(model_name, revision=revision)  # nosec B615
-            self.logger.info("BERT センチメント分析器初期化完了")
-        except ImportError:
-            self.logger.warning(
-                "transformersライブラリが利用不可 - 簡易センチメント分析を使用"
-            )
-            self.sentiment_analyzer = self._create_simple_sentiment_analyzer()
-        except Exception as e:
-            self.logger.error(f"BERT初期化エラー: {e}")
-            self.sentiment_analyzer = self._create_simple_sentiment_analyzer()
-
-    def _create_simple_sentiment_analyzer(self):
-        """簡易センチメント分析器"""
-        positive_words = ["上昇", "好調", "成長", "利益", "買い", "強気", "回復"]
-        negative_words = ["下落", "悪化", "減少", "損失", "売り", "弱気", "暴落"]
-
-        def analyze_sentiment(text: str) -> float:
-            pos_count = sum(1 for word in positive_words if word in text)
-            neg_count = sum(1 for word in negative_words if word in text)
-            if pos_count + neg_count == 0:
-                return 0.0
-            return (pos_count - neg_count) / (pos_count + neg_count)
-
-        return analyze_sentiment
-
-    def _initialize_macro_data(self):
-        """マクロ経済指標データ取得"""
-        try:
-            import yfinance as yf
-
-            # 主要指標取得
-            indicators = {
-                "usdjpy": "^USDJPY=X",  # ドル円
-                "nikkei": "^N225",  # 日経平均
-                "sp500": "^GSPC",  # S&P500
-                "vix": "^VIX",  # VIX恐怖指数
-                "dxy": "DX-Y.NYB",  # ドル指数
-            }
-            for name, symbol in indicators.items():
-                try:
-                    data = yf.download(symbol, period="1y", progress=False)
-                    if not data.empty:
-                        self.macro_data[name] = data["Close"].pct_change().fillna(0)
-                        self.logger.info(f"{name} マクロ指標データ取得完了")
-                except Exception as e:
-                    self.logger.warning(f"{name} データ取得失敗: {e}")
-        except Exception as e:
-            self.logger.error(f"マクロ経済データ初期化エラー: {e}")
-
-    def enhanced_sentiment_prediction(self, symbol: str) -> Dict[str, float]:
-        """強化センチメント予測"""
-        try:
-            from analysis.sentiment_analyzer import MarketSentimentAnalyzer
-
-            analyzer = MarketSentimentAnalyzer()
-            sentiment_result = analyzer.analyze_news_sentiment(symbol)
-            # dictから適切な値を取得
-            if isinstance(sentiment_result, dict):
-                sentiment_score = sentiment_result.get("sentiment_score", 0.0)
-                base_confidence = sentiment_result.get("confidence", 0.1)
-            else:
-                sentiment_score = float(sentiment_result) if sentiment_result else 0.0
-                base_confidence = 0.1
-            # BERT強化分析（利用可能な場合）
-            if hasattr(self, "bert_model"):
-                try:
-                    enhanced_score = self._bert_sentiment_analysis(symbol)
-                    # 重み付き統合
-                    final_score = 0.7 * sentiment_score + 0.3 * enhanced_score
-                    final_confidence = max(base_confidence, 0.5)
-                except Exception as bert_error:
-                    self.logger.error(f"BERT分析エラー: {bert_error}")
-                    final_score = sentiment_score
-                    final_confidence = base_confidence
-            else:
-                final_score = sentiment_score
-                final_confidence = base_confidence
-            return {
-                "sentiment_score": float(final_score),
-                "confidence": float(final_confidence),
-            }
-        except Exception as e:
-            self.logger.error(f"センチメント予測エラー {symbol}: {e}")
-            return {"sentiment_score": 0.0, "confidence": 0.0}
-
-    def _bert_sentiment_analysis(self, symbol: str) -> float:
-        """BERT活用高度センチメント分析"""
-        try:
-            # ニュースデータ取得（実装は analysis/sentiment_analyzer.py と連携）
-            import yfinance as yf
-
-            ticker = yf.Ticker(f"{symbol}.T")
-            news_data = ticker.news
-            if not news_data:
-                return 0.0
-            sentiments = []
-            for article in news_data[:5]:  # 最新5記事
-                title = article.get("title", "")
-                # BERT tokenize & predict
-                inputs = self.tokenizer(
-                    title,
-                    return_tensors="pt",
-                    truncation=True,
-                    padding=True,
-                    max_length=128,
-                )
-                with torch.no_grad():
-                    outputs = self.bert_model(**inputs)
-                    probabilities = torch.nn.functional.softmax(outputs.logits, dim=-1)
-                    # ポジティブ/ネガティブスコア算出
-                    sentiment = probabilities[0][1].item() - probabilities[0][0].item()
-                    sentiments.append(sentiment)
-            return np.mean(sentiments) if sentiments else 0.0
-        except Exception as e:
-            self.logger.error(f"BERT分析エラー: {e}")
-            return 0.0
-
-    def macro_enhanced_prediction(self, symbol: str) -> Dict[str, float]:
-        """マクロ経済指標強化予測"""
-        try:
-            # 銘柄の業界・特性分析
-            sector_sensitivity = self._analyze_sector_sensitivity(symbol)
-            # マクロ指標による調整
-            macro_adjustment = 0.0
-            for indicator, sensitivity in sector_sensitivity.items():
-                if indicator in self.macro_data:
-                    macro_value = self.macro_data[indicator]
-                    # Series型のデータを安全に処理
-                    if hasattr(macro_value, "tail") and len(macro_value) > 0:
-                        # pandasシリーズの場合
-                        recent_values = macro_value.tail(5)
-                        if len(recent_values) > 0:
-                            recent_change = float(recent_values.mean())
-                        else:
-                            recent_change = 0.0
-                    elif isinstance(macro_value, (int, float)):
-                        # 数値の場合
-                        recent_change = float(macro_value)
-                    else:
-                        # その他の場合はスキップ
-                        recent_change = 0.0
-                    macro_adjustment += recent_change * sensitivity
-            # スコア正規化
-            macro_score = np.tanh(macro_adjustment * 10)  # -1 to 1
-            return {
-                "macro_score": float(macro_score),
-                "confidence": float(min(abs(macro_score) * 2, 1.0)),
-            }
-        except Exception as e:
-            self.logger.error(f"マクロ予測エラー {symbol}: {e}")
-            return {"macro_score": 0.0, "confidence": 0.0}
-
-    def _analyze_sector_sensitivity(self, symbol: str) -> Dict[str, float]:
-        """業界別マクロ経済感応度分析"""
-        # 簡易版：銘柄コードベース業界推定
-        sector_map = {
-            # 自動車 (7000番台)
-            "7": {"usdjpy": 0.8, "sp500": 0.6, "vix": -0.4},
-            # 電機 (6000番台)
-            "6": {"usdjpy": 0.7, "sp500": 0.8, "vix": -0.6},
-            # 金融 (8000番台)
-            "8": {"usdjpy": 0.3, "nikkei": 0.9, "vix": -0.8},
-            # 通信 (9000番台)
-            "9": {"usdjpy": 0.2, "sp500": 0.4, "vix": -0.3},
-        }
-        first_digit = symbol[0] if symbol else "0"
-        return sector_map.get(first_digit, {"usdjpy": 0.5, "nikkei": 0.5, "vix": -0.5})
-
-    def dynamic_ensemble_prediction(self, symbol: str) -> Dict[str, Any]:
-        """動的重み調整アンサンブル予測"""
-        try:
-            predictions = {}
-            confidences = {}
-            # 1. 84.6%ベースモデル予測
-            base_pred = self._get_base_prediction(symbol)
-            predictions["trend_following"] = base_pred["prediction"]
-            confidences["trend_following"] = base_pred["confidence"]
-            # 2. 深層学習予測
-            if "lstm_deep" in self.models:
-                lstm_score = self.models["lstm_deep"].predict_deep(symbol)
-                predictions["lstm_deep"] = lstm_score
-                confidences["lstm_deep"] = 0.7  # 固定信頼度
-            # 3. センチメント強化予測
-            sentiment_result = self.enhanced_sentiment_prediction(symbol)
-            sentiment_pred = (
-                50 + sentiment_result["sentiment_score"] * 25
-            )  # -1~1 を 25~75に変換
-            predictions["sentiment_enhanced"] = sentiment_pred
-            confidences["sentiment_enhanced"] = sentiment_result["confidence"]
-            # 4. マクロ経済強化予測
-            macro_result = self.macro_enhanced_prediction(symbol)
-            macro_pred = 50 + macro_result["macro_score"] * 25
-            predictions["macro_enhanced"] = macro_pred
-            confidences["macro_enhanced"] = macro_result["confidence"]
-            # 5. 動的重み調整
-            adjusted_weights = self._adjust_weights_dynamically(confidences)
-            # 6. アンサンブル予測計算
-            ensemble_score = 0.0
-            total_weight = 0.0
-            for model_name, pred in predictions.items():
-                if pred is not None and model_name in adjusted_weights:
-                    weight = adjusted_weights[model_name] * confidences.get(
-                        model_name, 0.5
-                    )
-                    ensemble_score += pred * weight
-                    total_weight += weight
-            if total_weight > 0:
-                ensemble_score /= total_weight
-            else:
-                ensemble_score = 50.0  # デフォルト
-            # 信頼度算出
-            ensemble_confidence = min(
-                total_weight / sum(adjusted_weights.values()), 1.0
-            )
-            return {
-                "ensemble_prediction": ensemble_score,
-                "ensemble_confidence": ensemble_confidence,
-                "individual_predictions": predictions,
-                "adjusted_weights": adjusted_weights,
-                "high_confidence": ensemble_confidence >= self.confidence_threshold,
-            }
-        except Exception as e:
-            self.logger.error(f"アンサンブル予測エラー {symbol}: {e}")
-            return {
-                "ensemble_prediction": 50.0,
-                "ensemble_confidence": 0.0,
-                "individual_predictions": {},
-                "adjusted_weights": {},
-                "high_confidence": False,
-            }
-
-    def _get_base_prediction(self, symbol: str) -> Dict[str, float]:
-        """84.6%ベースモデル予測取得"""
-        try:
-            from trend_following_predictor import TrendFollowingPredictor
-
-            predictor = TrendFollowingPredictor()
-            result = predictor.predict_stock(symbol)
-            # 方向性を0-100スコアに変換
-            base_score = 75 if result["direction"] == 1 else 25
-            confidence = result["confidence"]
-            return {"prediction": base_score, "confidence": confidence}
-        except Exception as e:
-            self.logger.error(f"ベース予測エラー {symbol}: {e}")
-            return {"prediction": 50.0, "confidence": 0.0}
-
-    def _adjust_weights_dynamically(
-        self, confidences: Dict[str, float]
-    ) -> Dict[str, float]:
-        """信頼度ベース動的重み調整"""
-        adjusted_weights = {}
-        for model_name, base_weight in self.weights.items():
-            confidence = confidences.get(model_name, 0.5)
-            # 信頼度が高いモデルの重みを増加
-            if confidence >= 0.8:
-                adjustment = 1.5
-            elif confidence >= 0.6:
-                adjustment = 1.2
-            elif confidence >= 0.4:
-                adjustment = 1.0
-            else:
-                adjustment = 0.8
-            adjusted_weights[model_name] = base_weight * adjustment
-        # 正規化
-        total = sum(adjusted_weights.values())
-        if total > 0:
-            adjusted_weights = {k: v / total for k, v in adjusted_weights.items()}
-        return adjusted_weights
-
-    def train_ensemble_system(self, symbols: List[str]):
-        """アンサンブルシステム全体訓練"""
-        try:
-            self.logger.info("アンサンブル学習システム訓練開始")
-            # 各コンポーネント訓練
-            if "lstm_deep" in self.models:
-                self.logger.info("LSTM深層学習訓練中...")
-                self.models["lstm_deep"].train_deep_model(symbols[:10])  # 計算量制限
-            if "transformer_deep" in self.models:
-                self.logger.info("Transformer深層学習訓練中...")
-                self.models["transformer_deep"].train_deep_model(symbols[:10])
-            # 重み最適化（バックテストベース）
-            self._optimize_ensemble_weights(symbols[:20])
-            self.logger.info("アンサンブル学習システム訓練完了")
-        except Exception as e:
-            self.logger.error(f"アンサンブル訓練エラー: {e}")
-
-    def _optimize_ensemble_weights(self, symbols: List[str]):
-        """バックテストベース重み最適化"""
-        try:
-            from scipy.optimize import minimize
-
-            def objective(weights_array):
-                """最適化目的関数"""
-                # weights_arrayを辞書に変換
-                weight_names = list(self.weights.keys())
-                weights_dict = dict(zip(weight_names, weights_array))
-                total_accuracy = 0
-                valid_predictions = 0
-                for symbol in symbols[:10]:  # サンプル制限
-                    try:
-                        # 実際の予測と検証（簡易版）
-                        ensemble_result = self.dynamic_ensemble_prediction(symbol)
-                        if ensemble_result["high_confidence"]:
-                            # 簡易精度評価（実際はバックテストが必要）
-                            total_accuracy += ensemble_result["ensemble_confidence"]
-                            valid_predictions += 1
-                    except:
-                        continue
-                return -(total_accuracy / max(valid_predictions, 1))  # 最大化のため負値
-
-            # 初期重み
-            initial_weights = list(self.weights.values())
-            # 制約：重みの合計=1
-            constraints = {"type": "eq", "fun": lambda x: np.sum(x) - 1}
-            bounds = [(0.05, 0.6) for _ in initial_weights]  # 各重み5%-60%
-            # 最適化実行
-            result = minimize(
-                objective,
-                initial_weights,
-                method="SLSQP",
-                bounds=bounds,
-                constraints=constraints,
-            )
-            if result.success:
-                weight_names = list(self.weights.keys())
-                optimized_weights = dict(zip(weight_names, result.x))
-                self.weights = optimized_weights
-                self.logger.info(f"重み最適化完了: {optimized_weights}")
-        except Exception as e:
-            self.logger.error(f"重み最適化エラー: {e}")
-
-
-class MacroEconomicDataProvider:
-    """
-    マクロ経済指標データプロバイダー
-    日銀政策・金利・為替等の経済指標を統合管理
-    """
-
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
-        self.data_cache = {}
-        self.last_update = {}
-
-    def get_boj_policy_data(self) -> Dict[str, Any]:
-        """日本銀行政策データ取得"""
-        try:
-            # 実際のAPIがある場合の実装想定
-            # ここでは簡易版として固定値＋変動を返す
-            policy_data = {
-                "interest_rate": 0.1,  # 政策金利
-                "money_supply_growth": 2.5,  # マネーサプライ成長率
-                "inflation_target": 2.0,  # インフレ目標
-                "yield_curve_control": True,  # イールドカーブコントロール
-                "policy_stance": "accommodative",  # 政策スタンス
-                "last_meeting_date": "2025-01-23",
-                "next_meeting_date": "2025-03-19",
-            }
-            self.logger.info("日銀政策データ取得完了")
-            return policy_data
-        except Exception as e:
-            self.logger.error(f"日銀政策データ取得エラー: {e}")
-            return {}
-
-    def get_global_rates_data(self) -> Dict[str, float]:
-        """世界主要国金利データ取得"""
-        try:
-            import yfinance as yf
-
-            rates_symbols = {
-                "us_10y": "^TNX",  # 米10年債利回り
-                "jp_10y": "^TNX",  # 日10年債利回り（簡易）
-                "fed_rate": "^IRX",  # 米短期金利
-            }
-            rates_data = {}
-            for name, symbol in rates_symbols.items():
-                try:
-                    data = yf.download(symbol, period="5d", progress=False)
-                    if not data.empty:
-                        rates_data[name] = data["Close"].iloc[-1]
-                except:
-                    rates_data[name] = 0.0
-            return rates_data
-        except Exception as e:
-            self.logger.error(f"金利データ取得エラー: {e}")
-            return {}
-
-    def get_economic_indicators(self) -> Dict[str, Any]:
-        """総合経済指標取得"""
-        try:
-            indicators = {
-                "boj_policy": self.get_boj_policy_data(),
-                "global_rates": self.get_global_rates_data(),
-                "currency_strength": self._get_currency_strength(),
-                "market_sentiment": self._get_market_sentiment_indicators(),
-            }
-            return indicators
-        except Exception as e:
-            self.logger.error(f"経済指標取得エラー: {e}")
-            return {}
-
-    def _get_currency_strength(self) -> Dict[str, float]:
-        """通貨強度指標"""
-        try:
-            import yfinance as yf
-
-            # 主要通貨ペア
-            currency_pairs = {
-                "USDJPY": "USDJPY=X",
-                "EURJPY": "EURJPY=X",
-                "GBPJPY": "GBPJPY=X",
-            }
-            strength_data = {}
-            for pair, symbol in currency_pairs.items():
-                try:
-                    data = yf.download(symbol, period="1mo", progress=False)
-                    if not data.empty:
-                        # 1ヶ月変化率
-                        change = (
-                            data["Close"].iloc[-1] / data["Close"].iloc[0] - 1
-                        ) * 100
-                        strength_data[pair] = change
-                except:
-                    strength_data[pair] = 0.0
-            return strength_data
-        except Exception as e:
-            self.logger.error(f"通貨強度取得エラー: {e}")
-            return {}
-
-    def _get_market_sentiment_indicators(self) -> Dict[str, float]:
-        """市場センチメント指標"""
-        try:
-            import yfinance as yf
-
-            sentiment_symbols = {
-                "vix": "^VIX",  # VIX恐怖指数
-                "vix_jp": "^N225",  # 日経VI（簡易版）
-                "put_call_ratio": "^VIX",  # プット/コール比率（簡易）
-            }
-            sentiment_data = {}
-            for name, symbol in sentiment_symbols.items():
-                try:
-                    data = yf.download(symbol, period="1mo", progress=False)
-                    if (
-                        len(data) > 0 and "Close" in data.columns
-                    ):  # Series比較エラー修正
-                        current_val = data["Close"].iloc[-1]
-                        avg_val = data["Close"].mean()
-                        # 平均との乖離率
-                        deviation = (current_val / avg_val - 1) * 100
-                        sentiment_data[name] = float(deviation)
-                    else:
-                        sentiment_data[name] = 0.0
-                except Exception as symbol_error:
-                    self.logger.warning(
-                        f"センチメント指標取得エラー {name}: {symbol_error}"
-                    )
-                    sentiment_data[name] = 0.0
-            return sentiment_data
-        except Exception as e:
-            self.logger.error(f"センチメント指標取得エラー: {e}")
-            return {}
-
 
 class AdvancedPrecisionBreakthrough87System:
     """
@@ -2713,331 +884,6 @@ class AdvancedPrecisionBreakthrough87System:
             self.logger.error(f"バッチ予測エラー: {e}")
             return {"error": str(e), "total_symbols": len(symbols)}
 
-
-class MetaLearningOptimizer:
-    """メタ学習最適化システム - 銘柄特性への適応学習"""
-
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
-        self.symbol_profiles = {}
-        self.adaptation_memory = {}
-
-    def create_symbol_profile(
-        self, symbol: str, historical_data: pd.DataFrame
-    ) -> Dict[str, float]:
-        """銘柄特性プロファイル作成"""
-        try:
-            profile = {
-                "volatility_regime": self._analyze_volatility_regime(historical_data),
-                "trend_persistence": self._analyze_trend_persistence(historical_data),
-                "volume_pattern": self._analyze_volume_pattern(historical_data),
-                "sector_correlation": self._analyze_sector_correlation(symbol),
-                "liquidity_score": self._calculate_liquidity_score(historical_data),
-                "momentum_sensitivity": self._analyze_momentum_sensitivity(
-                    historical_data
-                ),
-            }
-            self.symbol_profiles[symbol] = profile
-            return profile
-        except Exception as e:
-            self.logger.error(f"銘柄プロファイル作成エラー {symbol}: {e}")
-            return {}
-
-    def _analyze_volatility_regime(self, data: pd.DataFrame) -> float:
-        """ボラティリティ体制分析"""
-        returns = data["Close"].pct_change().dropna()
-        # GARCH的ボラティリティクラスタリング検出
-        volatility = returns.rolling(20).std()
-        vol_changes = volatility.diff().abs()
-        # 高ボラティリティ期間の持続性
-        high_vol_periods = (volatility > volatility.quantile(0.8)).astype(int)
-        persistence = high_vol_periods.rolling(10).sum().mean() / 10
-        return float(persistence)
-
-    def _analyze_trend_persistence(self, data: pd.DataFrame) -> float:
-        """トレンド持続性分析"""
-        prices = data["Close"]
-        # 複数時間軸でのトレンド一貫性
-        sma_5 = prices.rolling(5).mean()
-        sma_20 = prices.rolling(20).mean()
-        sma_60 = prices.rolling(60).mean()
-        # トレンド方向の一致度
-        trend_5_20 = (sma_5 > sma_20).astype(int)
-        trend_20_60 = (sma_20 > sma_60).astype(int)
-        consistency = (trend_5_20 == trend_20_60).mean()
-        return float(consistency)
-
-    def _analyze_volume_pattern(self, data: pd.DataFrame) -> float:
-        """出来高パターン分析"""
-        volume = data["Volume"]
-        price_change = data["Close"].pct_change().abs()
-        # 価格変動と出来高の相関
-        correlation = price_change.corr(volume)
-        return float(correlation) if not np.isnan(correlation) else 0.5
-
-    def _analyze_sector_correlation(self, symbol: str) -> float:
-        """セクター相関分析"""
-        # 簡易セクター分類による相関スコア
-        tech_stocks = ["6758", "9984", "4519"]
-        finance_stocks = ["8306", "8035"]
-        if symbol in tech_stocks:
-            return 0.8  # 高相関
-        elif symbol in finance_stocks:
-            return 0.7
-        else:
-            return 0.5  # 中立
-
-    def _calculate_liquidity_score(self, data: pd.DataFrame) -> float:
-        """流動性スコア計算"""
-        volume = data["Volume"]
-        avg_volume = volume.mean()
-        volume_stability = 1.0 - (volume.std() / avg_volume)
-        return float(np.clip(volume_stability, 0.0, 1.0))
-
-    def _analyze_momentum_sensitivity(self, data: pd.DataFrame) -> float:
-        """モメンタム感応度分析"""
-        returns = data["Close"].pct_change()
-        # 短期モメンタムの価格反応
-        momentum_3 = returns.rolling(3).sum()
-        momentum_10 = returns.rolling(10).sum()
-        # モメンタム継続性
-        momentum_persistence = (momentum_3 * momentum_10 > 0).mean()
-        return float(momentum_persistence)
-
-    def adapt_model_parameters(
-        self, symbol: str, base_prediction: float, confidence: float
-    ) -> Dict[str, float]:
-        """モデルパラメータの適応調整"""
-        try:
-            if symbol not in self.symbol_profiles:
-                return {
-                    "adapted_prediction": base_prediction,
-                    "adapted_confidence": confidence,
-                }
-            profile = self.symbol_profiles[symbol]
-            # プロファイルベースの調整
-            volatility_adjustment = (profile["volatility_regime"] - 0.5) * 0.1
-            trend_adjustment = (profile["trend_persistence"] - 0.5) * 0.2
-            momentum_adjustment = (profile["momentum_sensitivity"] - 0.5) * 0.15
-            # 予測値調整
-            adapted_prediction = (
-                base_prediction
-                + (volatility_adjustment + trend_adjustment + momentum_adjustment) * 10
-            )
-            # 信頼度調整
-            confidence_boost = (
-                profile["trend_persistence"] * profile["liquidity_score"] * 0.3
-            )
-            adapted_confidence = min(confidence + confidence_boost, 1.0)
-            return {
-                "adapted_prediction": float(adapted_prediction),
-                "adapted_confidence": float(adapted_confidence),
-                "adjustments": {
-                    "volatility": volatility_adjustment,
-                    "trend": trend_adjustment,
-                    "momentum": momentum_adjustment,
-                },
-            }
-        except Exception as e:
-            self.logger.error(f"パラメータ適応エラー {symbol}: {e}")
-            return {
-                "adapted_prediction": base_prediction,
-                "adapted_confidence": confidence,
-            }
-
-
-class DQNReinforcementLearner:
-    """DQN強化学習システム - 市場環境への動的適応"""
-
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
-        self.q_network = self._build_q_network()
-        self.target_network = self._build_q_network()
-        self.memory = []
-        self.epsilon = 0.3  # 探索率
-        self.gamma = 0.95  # 割引率
-        self.learning_rate = 0.001
-
-    def _build_q_network(self) -> Dict[str, Any]:
-        """Q-ネットワーク構築"""
-        return {
-            "input_dim": 15,  # 市場状態次元
-            "hidden_dims": [64, 32, 16],
-            "output_dim": 3,  # アクション: [買い, 保持, 売り]
-            "weights": self._initialize_weights(),
-        }
-
-    def _initialize_weights(self) -> Dict[str, np.ndarray]:
-        """ネットワーク重み初期化"""
-        return {
-            "w1": np.random.randn(15, 64) * 0.1,
-            "b1": np.zeros(64),
-            "w2": np.random.randn(64, 32) * 0.1,
-            "b2": np.zeros(32),
-            "w3": np.random.randn(32, 16) * 0.1,
-            "b3": np.zeros(16),
-            "w4": np.random.randn(16, 3) * 0.1,
-            "b4": np.zeros(3),
-        }
-
-    def extract_market_state(
-        self, symbol: str, historical_data: pd.DataFrame
-    ) -> np.ndarray:
-        """市場状態特徴量抽出"""
-        try:
-            if len(historical_data) < 50:
-                return np.zeros(15)
-            data = historical_data.tail(50).copy()
-            # 価格系特徴量
-            returns = data["Close"].pct_change().dropna()
-            volatility = returns.rolling(10).std().iloc[-1]
-            momentum = returns.tail(5).mean()
-            # テクニカル指標
-            sma_5 = data["Close"].rolling(5).mean().iloc[-1]
-            sma_20 = data["Close"].rolling(20).mean().iloc[-1]
-            current_price = data["Close"].iloc[-1]
-            # RSI
-            delta = data["Close"].diff()
-            gain = delta.where(delta > 0, 0).rolling(14).mean().iloc[-1]
-            loss = (-delta.where(delta < 0, 0)).rolling(14).mean().iloc[-1]
-            rsi = 100 - (100 / (1 + gain / loss)) if loss != 0 else 50
-            # MACD
-            ema_12 = data["Close"].ewm(span=12).mean().iloc[-1]
-            ema_26 = data["Close"].ewm(span=26).mean().iloc[-1]
-            macd = ema_12 - ema_26
-            # 出来高系
-            volume_ratio = data["Volume"].iloc[-1] / data["Volume"].tail(20).mean()
-            # 市場状態ベクター
-            state = np.array(
-                [
-                    volatility,
-                    momentum,
-                    (current_price - sma_5) / sma_5,
-                    (current_price - sma_20) / sma_20,
-                    (sma_5 - sma_20) / sma_20,
-                    rsi / 100,
-                    macd,
-                    volume_ratio,
-                    returns.tail(1).iloc[0],
-                    returns.tail(3).mean(),
-                    returns.tail(10).mean(),
-                    returns.tail(10).std(),
-                    (data["High"].iloc[-1] - data["Low"].iloc[-1])
-                    / data["Close"].iloc[-1],
-                    data["Close"].iloc[-1] / data["Close"].iloc[-5] - 1,
-                    len(data),
-                ]
-            )
-            # NaN値処理
-            state = np.nan_to_num(state, 0.0)
-            return state.astype(np.float32)
-        except Exception as e:
-            self.logger.error(f"市場状態抽出エラー {symbol}: {e}")
-            return np.zeros(15, dtype=np.float32)
-
-    def forward_pass(self, state: np.ndarray, network: Dict[str, Any]) -> np.ndarray:
-        """フォワードパス"""
-        try:
-            weights = network["weights"]
-            # Layer 1
-            z1 = np.dot(state, weights["w1"]) + weights["b1"]
-            a1 = np.maximum(0, z1)  # ReLU
-            # Layer 2
-            z2 = np.dot(a1, weights["w2"]) + weights["b2"]
-            a2 = np.maximum(0, z2)  # ReLU
-            # Layer 3
-            z3 = np.dot(a2, weights["w3"]) + weights["b3"]
-            a3 = np.maximum(0, z3)  # ReLU
-            # Output Layer
-            q_values = np.dot(a3, weights["w4"]) + weights["b4"]
-            return q_values
-        except Exception as e:
-            self.logger.error(f"フォワードパスエラー: {e}")
-            return np.array([0.5, 0.5, 0.5])
-
-    def select_action(self, state: np.ndarray) -> int:
-        """行動選択 (ε-greedy)"""
-        if np.random.random() < self.epsilon:
-            return np.random.randint(3)  # ランダム行動
-        q_values = self.forward_pass(state, self.q_network)
-        return np.argmax(q_values)
-
-    def get_trading_signal(
-        self, symbol: str, historical_data: pd.DataFrame
-    ) -> Dict[str, Any]:
-        """取引シグナル生成 - 87%精度向上版"""
-        try:
-            state = self.extract_market_state(symbol, historical_data)
-            action = self.select_action(state)
-            q_values = self.forward_pass(state, self.q_network)
-
-            # アクション解釈（強化版）
-            action_map = {0: "BUY", 1: "HOLD", 2: "SELL"}
-
-            # より強力なシグナル強度計算
-            q_max = float(np.max(q_values))
-            q_mean = float(np.mean(q_values))
-            q_std = float(np.std(q_values))
-
-            # 強化されたシグナル強度
-            signal_strength = (q_max - q_mean) + q_std * 0.5
-
-            # 市場状況による信頼度調整
-            market_volatility = float(np.std(state[:5])) if len(state) >= 5 else 0.1
-            trend_strength = (
-                float(abs(np.mean(state[5:10]))) if len(state) >= 10 else 0.5
-            )
-
-            # DQN信頼度の強化計算
-            base_confidence = float(q_max)
-            volatility_adjustment = min(
-                market_volatility * 2, 0.2
-            )  # ボラティリティボーナス
-            trend_adjustment = min(trend_strength * 0.3, 0.15)  # トレンド強度ボーナス
-
-            enhanced_confidence = min(
-                base_confidence + volatility_adjustment + trend_adjustment, 0.95
-            )
-
-            # アクション別の追加調整
-            if action == 0:  # BUY
-                signal_strength *= 1.2  # 買いシグナルを強化
-                if q_values[0] > 0.7:  # 強い買いシグナル
-                    enhanced_confidence *= 1.1
-            elif action == 2:  # SELL
-                signal_strength *= 1.1  # 売りシグナルを強化
-                if q_values[2] > 0.7:  # 強い売りシグナル
-                    enhanced_confidence *= 1.05
-
-            # シグナル強度の正規化と強化
-            signal_strength = float(np.clip(signal_strength * 1.5, -1.0, 1.0))
-
-            return {
-                "action": action_map[action],
-                "signal_strength": signal_strength,
-                "confidence": float(np.clip(enhanced_confidence, 0.3, 0.95)),
-                "q_values": q_values.tolist(),
-                "market_state": state.tolist(),
-                "enhancement_applied": {
-                    "volatility_adjustment": volatility_adjustment,
-                    "trend_adjustment": trend_adjustment,
-                    "signal_multiplier": 1.5,
-                },
-            }
-
-        except Exception as e:
-            self.logger.error(f"取引シグナル生成エラー {symbol}: {e}")
-            # フォールバック時も改善
-            return {
-                "action": "HOLD",
-                "signal_strength": 0.1,  # 少し改善
-                "confidence": 0.55,  # 少し改善
-                "q_values": [0.55, 0.5, 0.45],
-                "market_state": [],
-                "enhancement_applied": {"fallback": True},
-            }
-
-
 class Precision87BreakthroughSystem:
     """87%精度突破統合システム"""
 
@@ -3327,440 +1173,6 @@ class Precision87BreakthroughSystem:
             "component_breakdown": {},
             "tuning_applied": {},
         }
-
-
-class SentimentAnalyzer:
-    """センチメント分析器"""
-
-    def __init__(self):
-        self.sentiment_cache = {}
-
-    def get_news_sentiment(self, symbol: str) -> Dict[str, float]:
-        """ニュースセンチメント取得（模擬実装）"""
-        # 実際の実装では Yahoo Finance APIやNews APIを使用
-        import random
-
-        cache_key = f"sentiment_{symbol}_{datetime.now().strftime('%Y-%m-%d')}"
-        if cache_key in self.sentiment_cache:
-            return self.sentiment_cache[cache_key]
-        # 模擬センチメントデータ
-        sentiment_data = {
-            "positive_ratio": random.uniform(0.2, 0.8),  # nosec B311
-            "negative_ratio": random.uniform(0.1, 0.4),  # nosec B311
-            "neutral_ratio": random.uniform(0.2, 0.5),  # nosec B311
-            "news_volume": random.randint(5, 50),  # nosec B311
-            "sentiment_trend": random.uniform(-0.3, 0.3),  # nosec B311
-            "social_media_buzz": random.uniform(0.1, 0.9),  # nosec B311
-        }
-        # 正規化
-        total = (
-            sentiment_data["positive_ratio"]
-            + sentiment_data["negative_ratio"]
-            + sentiment_data["neutral_ratio"]
-        )
-        for key in ["positive_ratio", "negative_ratio", "neutral_ratio"]:
-            sentiment_data[key] /= total
-        self.sentiment_cache[cache_key] = sentiment_data
-        return sentiment_data
-
-    def get_macro_economic_features(self) -> Dict[str, float]:
-        """マクロ経済指標取得（模擬実装）"""
-        # 実際の実装では FRED API や日本銀行 API を使用
-        import random
-
-        return {
-            "interest_rate": random.uniform(0.001, 0.05),  # nosec B311
-            "inflation_rate": random.uniform(-0.01, 0.03),  # nosec B311
-            "gdp_growth": random.uniform(-0.02, 0.04),  # nosec B311
-            "unemployment_rate": random.uniform(0.02, 0.06),  # nosec B311
-            "exchange_rate_usd_jpy": random.uniform(140, 160),  # nosec B311
-            "oil_price": random.uniform(70, 120),  # nosec B311
-            "gold_price": random.uniform(1800, 2200),  # nosec B311
-            "vix_index": random.uniform(10, 40),  # nosec B311
-            "nikkei_momentum": random.uniform(-0.05, 0.05),  # nosec B311
-        }
-
-
-class RedisCache:
-    """Redis高速キャッシュシステム"""
-
-    def __init__(self, host="localhost", port=6379, db=0):
-        try:
-            import redis
-
-            self.redis_client = redis.Redis(
-                host=host, port=port, db=db, decode_responses=True, socket_timeout=5
-            )
-            # 接続テスト
-            self.redis_client.ping()
-            self.redis_available = True
-            logger.info("Redis cache connected successfully")
-        except Exception as e:
-            logger.warning(
-                f"Redis not available, falling back to memory cache: {str(e)}"
-            )
-            self.redis_available = False
-            self.memory_cache = {}
-
-    def get(self, key: str) -> Optional[str]:
-        """キャッシュ取得"""
-        try:
-            if self.redis_available:
-                return self.redis_client.get(key)
-            else:
-                return self.memory_cache.get(key)
-        except Exception as e:
-            logger.error(f"Cache get error: {str(e)}")
-            return None
-
-    def set(self, key: str, value: str, ttl: int = 3600):
-        """キャッシュ設定"""
-        try:
-            if self.redis_available:
-                self.redis_client.setex(key, ttl, value)
-            else:
-                self.memory_cache[key] = value
-                # メモリキャッシュのTTL管理は簡略化
-        except Exception as e:
-            logger.error(f"Cache set error: {str(e)}")
-
-    def get_json(self, key: str) -> Optional[Dict]:
-        """JSON形式でキャッシュ取得"""
-        try:
-            data = self.get(key)
-            if data:
-                import json
-
-                return json.loads(data)
-            return None
-        except Exception as e:
-            logger.error(f"Cache JSON get error: {str(e)}")
-            return None
-
-    def set_json(self, key: str, value: Dict, ttl: int = 3600):
-        """JSON形式でキャッシュ設定"""
-        try:
-            import json
-
-            self.set(key, json.dumps(value), ttl)
-        except Exception as e:
-            logger.error(f"Cache JSON set error: {str(e)}")
-
-
-class MetaLearningOptimizer:
-    """メタラーニング・自動モデル選択システム"""
-
-    def __init__(self):
-        self.model_performance_history = {}
-        self.meta_features = {}
-        self.best_model_for_symbol = {}
-
-    def extract_meta_features(
-        self, symbol: str, data: pd.DataFrame
-    ) -> Dict[str, float]:
-        """メタ特徴量抽出"""
-        if data.empty:
-            return {}
-        meta_features = {
-            # データ特性
-            "data_length": len(data),
-            "missing_ratio": data.isnull().sum().sum()
-            / (data.shape[0] * data.shape[1]),
-            # 価格特性
-            "price_volatility": data["Close"].std() / data["Close"].mean(),
-            "price_trend": (data["Close"].iloc[-1] - data["Close"].iloc[0])
-            / data["Close"].iloc[0],
-            "price_skewness": data["Close"].skew(),
-            "price_kurtosis": data["Close"].kurtosis(),
-            # ボリューム特性
-            "volume_volatility": data["Volume"].std() / data["Volume"].mean(),
-            "volume_trend": (
-                data["Volume"].iloc[-20:].mean() - data["Volume"].iloc[:20].mean()
-            )
-            / data["Volume"].iloc[:20].mean(),
-            # 技術指標特性
-            "rsi_avg": data.get("RSI", pd.Series([50])).mean(),
-            "macd_trend": data.get("MACD", pd.Series([0])).tail(10).mean(),
-            # 市場特性
-            "sector_correlation": 0.5,  # 業界相関（後で実装）
-            "market_cap_category": 1.0,  # 時価総額カテゴリ（後で実装）
-        }
-        return meta_features
-
-    def select_best_model(self, symbol: str, data: pd.DataFrame) -> str:
-        """銘柄に最適なモデルを選択"""
-        meta_features = self.extract_meta_features(symbol, data)
-        # 過去の性能履歴から最適モデル選択
-        if symbol in self.best_model_for_symbol:
-            return self.best_model_for_symbol[symbol]
-        # メタ特徴量に基づく推奨
-        if meta_features.get("price_volatility", 0) > 0.05:
-            return "ensemble"  # 高ボラティリティ → アンサンブル
-        elif meta_features.get("data_length", 0) > 500:
-            return "deep_learning"  # 長期データ → 深層学習
-        else:
-            return "xgboost"  # デフォルト → XGBoost
-
-    def update_model_performance(
-        self, symbol: str, model_name: str, performance: float
-    ):
-        """モデル性能を更新"""
-        if symbol not in self.model_performance_history:
-            self.model_performance_history[symbol] = {}
-        if model_name not in self.model_performance_history[symbol]:
-            self.model_performance_history[symbol][model_name] = []
-        self.model_performance_history[symbol][model_name].append(
-            {"timestamp": datetime.now(), "performance": performance}
-        )
-        # 最新10回の平均性能で最適モデル更新
-        recent_performances = {}
-        for model, history in self.model_performance_history[symbol].items():
-            recent = [h["performance"] for h in history[-10:]]
-            recent_performances[model] = np.mean(recent) if recent else 0
-        if recent_performances:
-            self.best_model_for_symbol[symbol] = max(
-                recent_performances.keys(), key=lambda k: recent_performances[k]
-            )
-
-    def create_symbol_profile(self, symbol: str, data: pd.DataFrame) -> Dict[str, Any]:
-        """銘柄プロファイル作成 - 現在価格付き"""
-        try:
-            if data.empty or len(data) < 30:
-                return {
-                    "volatility_regime": "normal",
-                    "trend_persistence": 0.5,
-                    "volume_pattern": "stable",
-                    "price_momentum": 0.0,
-                    "seasonal_factor": 1.0,
-                    "sector_strength": 0.5,
-                    "current_price": (
-                        data["Close"].iloc[-1] if not data.empty else 100.0
-                    ),
-                }
-            close = data["Close"]
-            volume = data["Volume"]
-
-            # 現在価格を保存
-            current_price = float(close.iloc[-1])
-
-            # ボラティリティ体制分析
-            volatility = close.pct_change().rolling(20).std()
-            avg_vol = volatility.mean()
-            if avg_vol > 0.03:
-                volatility_regime = "high"
-            elif avg_vol < 0.015:
-                volatility_regime = "low"
-            else:
-                volatility_regime = "normal"
-
-            # トレンド持続性分析
-            returns = close.pct_change()
-            trend_periods = []
-            current_trend = 0
-            for ret in returns:
-                if pd.isna(ret):
-                    continue
-                if ret > 0.005:  # 0.5%以上上昇
-                    current_trend = current_trend + 1 if current_trend > 0 else 1
-                elif ret < -0.005:  # 0.5%以上下落
-                    current_trend = current_trend - 1 if current_trend < 0 else -1
-                else:
-                    if abs(current_trend) >= 3:
-                        trend_periods.append(abs(current_trend))
-                    current_trend = 0
-
-            trend_persistence = np.mean(trend_periods) / 10.0 if trend_periods else 0.5
-            trend_persistence = min(max(trend_persistence, 0.0), 1.0)
-
-            # ボリュームパターン分析
-            volume_sma = volume.rolling(20).mean()
-            volume_ratio = (
-                volume.iloc[-10:].mean() / volume_sma.iloc[-1]
-                if volume_sma.iloc[-1] > 0
-                else 1.0
-            )
-
-            if volume_ratio > 1.2:
-                volume_pattern = "increasing"
-            elif volume_ratio < 0.8:
-                volume_pattern = "decreasing"
-            else:
-                volume_pattern = "stable"
-
-            # 価格モメンタム
-            price_momentum = (
-                (close.iloc[-1] - close.iloc[-20]) / close.iloc[-20]
-                if len(close) >= 20
-                else 0.0
-            )
-
-            # 季節要因（簡易）
-            seasonal_factor = 1.0 + 0.1 * np.sin(
-                2 * np.pi * (pd.Timestamp.now().month - 1) / 12
-            )
-
-            # セクター強度（銘柄コードから推定）
-            first_digit = int(symbol[0]) if symbol and symbol[0].isdigit() else 5
-            sector_strength = 0.3 + (first_digit % 5) * 0.1  # 0.3-0.7の範囲
-
-            return {
-                "volatility_regime": volatility_regime,
-                "trend_persistence": trend_persistence,
-                "volume_pattern": volume_pattern,
-                "price_momentum": price_momentum,
-                "seasonal_factor": seasonal_factor,
-                "sector_strength": sector_strength,
-                "avg_volatility": avg_vol,
-                "volume_ratio": volume_ratio,
-                "current_price": current_price,  # 現在価格を追加
-            }
-
-        except Exception as e:
-            logger.error(f"Symbol profile creation error for {symbol}: {e}")
-            return {
-                "volatility_regime": "normal",
-                "trend_persistence": 0.5,
-                "volume_pattern": "stable",
-                "price_momentum": 0.0,
-                "seasonal_factor": 1.0,
-                "sector_strength": 0.5,
-                "current_price": 100.0,
-                "error": str(e),
-            }
-
-    def adapt_model_parameters(
-        self, symbol: str, symbol_profile: Dict[str, Any], base_params: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        """銘柄特性に基づくモデルパラメータ適応 - 87%精度向上版"""
-        try:
-            adapted_params = base_params.copy()
-
-            # ベース予測の改善
-            base_prediction = base_params.get("prediction", 50.0)
-            base_confidence = base_params.get("confidence", 0.5)
-
-            # ボラティリティ体制に基づく適応（強化版）
-            volatility_regime = symbol_profile.get("volatility_regime", "normal")
-            volatility_boost = 0
-
-            if volatility_regime == "high":
-                # 高ボラティリティ環境 - より慎重な予測
-                adapted_params["learning_rate"] = (
-                    base_params.get("learning_rate", 0.01) * 0.7
-                )
-                adapted_params["regularization"] = (
-                    base_params.get("regularization", 0.01) * 1.8
-                )
-                adapted_params["ensemble_diversity"] = 1.4
-                volatility_boost = -2.0  # 慎重な調整
-
-            elif volatility_regime == "low":
-                # 低ボラティリティ環境 - より積極的な予測
-                adapted_params["learning_rate"] = (
-                    base_params.get("learning_rate", 0.01) * 1.4
-                )
-                adapted_params["regularization"] = (
-                    base_params.get("regularization", 0.01) * 0.5
-                )
-                adapted_params["ensemble_diversity"] = 0.6
-                volatility_boost = 3.0  # 積極的な調整
-
-            else:
-                # 通常環境
-                adapted_params["ensemble_diversity"] = 1.0
-                volatility_boost = 1.0
-
-            # トレンド持続性に基づく適応（強化版）
-            trend_persistence = symbol_profile.get("trend_persistence", 0.5)
-            trend_boost = 0
-
-            if trend_persistence > 0.7:
-                # 強いトレンド持続性
-                adapted_params["momentum_factor"] = 1.5
-                adapted_params["trend_weight"] = 1.4
-                trend_boost = 4.0  # 強いトレンドを活用
-
-            elif trend_persistence < 0.3:
-                # 弱いトレンド持続性
-                adapted_params["momentum_factor"] = 0.6
-                adapted_params["trend_weight"] = 0.7
-                trend_boost = -1.0  # 反転を期待
-
-            else:
-                adapted_params["momentum_factor"] = 1.0
-                adapted_params["trend_weight"] = 1.0
-                trend_boost = 1.5
-
-            # ボリュームパターンに基づく適応（強化版）
-            volume_pattern = symbol_profile.get("volume_pattern", "stable")
-            volume_boost = 0
-
-            if volume_pattern == "increasing":
-                adapted_params["volume_weight"] = 1.5
-                volume_boost = 2.5  # 出来高増加は好材料
-
-            elif volume_pattern == "decreasing":
-                adapted_params["volume_weight"] = 0.6
-                volume_boost = -1.5  # 出来高減少は懸念材料
-
-            else:
-                adapted_params["volume_weight"] = 1.0
-                volume_boost = 0.5
-
-            # セクター強度に基づく適応（強化版）
-            sector_strength = symbol_profile.get("sector_strength", 0.5)
-            sector_boost = (sector_strength - 0.5) * 6  # より大きな影響
-            adapted_params["sector_adjustment"] = (
-                0.7 + sector_strength * 0.6
-            )  # 0.7-1.3の範囲
-
-            # 季節要因適応（強化版）
-            seasonal_factor = symbol_profile.get("seasonal_factor", 1.0)
-            seasonal_boost = (seasonal_factor - 1.0) * 3
-            adapted_params["seasonal_weight"] = seasonal_factor
-
-            # 総合適応予測計算
-            total_boost = (
-                volatility_boost
-                + trend_boost
-                + volume_boost
-                + sector_boost
-                + seasonal_boost
-            )
-            adapted_prediction = base_prediction + total_boost
-
-            # 適応信頼度計算
-            adaptation_strength = (
-                abs(total_boost) / 10.0
-            )  # ブースト強度から信頼度を算出
-            adapted_confidence = min(base_confidence + adaptation_strength * 0.1, 0.9)
-
-            # 最終パラメータ設定
-            adapted_params["adapted_prediction"] = float(
-                np.clip(adapted_prediction, 10, 90)
-            )
-            adapted_params["adapted_confidence"] = float(adapted_confidence)
-            adapted_params["meta_boost_applied"] = float(total_boost)
-            adapted_params["adaptation_details"] = {
-                "volatility_boost": volatility_boost,
-                "trend_boost": trend_boost,
-                "volume_boost": volume_boost,
-                "sector_boost": sector_boost,
-                "seasonal_boost": seasonal_boost,
-            }
-
-            return adapted_params
-
-        except Exception as e:
-            logger.error(f"Model parameter adaptation error for {symbol}: {e}")
-            # フォールバック時も少し改善
-            fallback = base_params.copy()
-            fallback["adapted_prediction"] = base_params.get("prediction", 50.0) + 1.0
-            fallback["adapted_confidence"] = min(
-                base_params.get("confidence", 0.5) + 0.05, 0.9
-            )
-            return fallback
-
 
 class UltraHighPerformancePredictor:
     """超高性能予測システム統合"""
