+++ models/sentiment.py
@@ -1,0 +1,188 @@
+"""Sentiment and macroeconomic analysis helpers."""
+
+import logging
+from datetime import datetime
+from typing import Any, Dict
+
+logger = logging.getLogger(__name__)
+
+class SentimentAnalyzer:
+    """センチメント分析器"""
+
+    def __init__(self):
+        self.sentiment_cache = {}
+
+    def get_news_sentiment(self, symbol: str) -> Dict[str, float]:
+        """ニュースセンチメント取得（模擬実装）"""
+        # 実際の実装では Yahoo Finance APIやNews APIを使用
+        import random
+
+        cache_key = f"sentiment_{symbol}_{datetime.now().strftime('%Y-%m-%d')}"
+        if cache_key in self.sentiment_cache:
+            return self.sentiment_cache[cache_key]
+        # 模擬センチメントデータ
+        sentiment_data = {
+            "positive_ratio": random.uniform(0.2, 0.8),  # nosec B311
+            "negative_ratio": random.uniform(0.1, 0.4),  # nosec B311
+            "neutral_ratio": random.uniform(0.2, 0.5),  # nosec B311
+            "news_volume": random.randint(5, 50),  # nosec B311
+            "sentiment_trend": random.uniform(-0.3, 0.3),  # nosec B311
+            "social_media_buzz": random.uniform(0.1, 0.9),  # nosec B311
+        }
+        # 正規化
+        total = (
+            sentiment_data["positive_ratio"]
+            + sentiment_data["negative_ratio"]
+            + sentiment_data["neutral_ratio"]
+        )
+        for key in ["positive_ratio", "negative_ratio", "neutral_ratio"]:
+            sentiment_data[key] /= total
+        self.sentiment_cache[cache_key] = sentiment_data
+        return sentiment_data
+
+    def get_macro_economic_features(self) -> Dict[str, float]:
+        """マクロ経済指標取得（模擬実装）"""
+        # 実際の実装では FRED API や日本銀行 API を使用
+        import random
+
+        return {
+            "interest_rate": random.uniform(0.001, 0.05),  # nosec B311
+            "inflation_rate": random.uniform(-0.01, 0.03),  # nosec B311
+            "gdp_growth": random.uniform(-0.02, 0.04),  # nosec B311
+            "unemployment_rate": random.uniform(0.02, 0.06),  # nosec B311
+            "exchange_rate_usd_jpy": random.uniform(140, 160),  # nosec B311
+            "oil_price": random.uniform(70, 120),  # nosec B311
+            "gold_price": random.uniform(1800, 2200),  # nosec B311
+            "vix_index": random.uniform(10, 40),  # nosec B311
+            "nikkei_momentum": random.uniform(-0.05, 0.05),  # nosec B311
+        }
+
+class MacroEconomicDataProvider:
+    """
+    マクロ経済指標データプロバイダー
+    日銀政策・金利・為替等の経済指標を統合管理
+    """
+
+    def __init__(self):
+        self.logger = logging.getLogger(__name__)
+        self.data_cache = {}
+        self.last_update = {}
+
+    def get_boj_policy_data(self) -> Dict[str, Any]:
+        """日本銀行政策データ取得"""
+        try:
+            # 実際のAPIがある場合の実装想定
+            # ここでは簡易版として固定値＋変動を返す
+            policy_data = {
+                "interest_rate": 0.1,  # 政策金利
+                "money_supply_growth": 2.5,  # マネーサプライ成長率
+                "inflation_target": 2.0,  # インフレ目標
+                "yield_curve_control": True,  # イールドカーブコントロール
+                "policy_stance": "accommodative",  # 政策スタンス
+                "last_meeting_date": "2025-01-23",
+                "next_meeting_date": "2025-03-19",
+            }
+            self.logger.info("日銀政策データ取得完了")
+            return policy_data
+        except Exception as e:
+            self.logger.error(f"日銀政策データ取得エラー: {e}")
+            return {}
+
+    def get_global_rates_data(self) -> Dict[str, float]:
+        """世界主要国金利データ取得"""
+        try:
+            import yfinance as yf
+
+            rates_symbols = {
+                "us_10y": "^TNX",  # 米10年債利回り
+                "jp_10y": "^TNX",  # 日10年債利回り（簡易）
+                "fed_rate": "^IRX",  # 米短期金利
+            }
+            rates_data = {}
+            for name, symbol in rates_symbols.items():
+                try:
+                    data = yf.download(symbol, period="5d", progress=False)
+                    if not data.empty:
+                        rates_data[name] = data["Close"].iloc[-1]
+                except:
+                    rates_data[name] = 0.0
+            return rates_data
+        except Exception as e:
+            self.logger.error(f"金利データ取得エラー: {e}")
+            return {}
+
+    def get_economic_indicators(self) -> Dict[str, Any]:
+        """総合経済指標取得"""
+        try:
+            indicators = {
+                "boj_policy": self.get_boj_policy_data(),
+                "global_rates": self.get_global_rates_data(),
+                "currency_strength": self._get_currency_strength(),
+                "market_sentiment": self._get_market_sentiment_indicators(),
+            }
+            return indicators
+        except Exception as e:
+            self.logger.error(f"経済指標取得エラー: {e}")
+            return {}
+
+    def _get_currency_strength(self) -> Dict[str, float]:
+        """通貨強度指標"""
+        try:
+            import yfinance as yf
+
+            # 主要通貨ペア
+            currency_pairs = {
+                "USDJPY": "USDJPY=X",
+                "EURJPY": "EURJPY=X",
+                "GBPJPY": "GBPJPY=X",
+            }
+            strength_data = {}
+            for pair, symbol in currency_pairs.items():
+                try:
+                    data = yf.download(symbol, period="1mo", progress=False)
+                    if not data.empty:
+                        # 1ヶ月変化率
+                        change = (
+                            data["Close"].iloc[-1] / data["Close"].iloc[0] - 1
+                        ) * 100
+                        strength_data[pair] = change
+                except:
+                    strength_data[pair] = 0.0
+            return strength_data
+        except Exception as e:
+            self.logger.error(f"通貨強度取得エラー: {e}")
+            return {}
+
+    def _get_market_sentiment_indicators(self) -> Dict[str, float]:
+        """市場センチメント指標"""
+        try:
+            import yfinance as yf
+
+            sentiment_symbols = {
+                "vix": "^VIX",  # VIX恐怖指数
+                "vix_jp": "^N225",  # 日経VI（簡易版）
+                "put_call_ratio": "^VIX",  # プット/コール比率（簡易）
+            }
+            sentiment_data = {}
+            for name, symbol in sentiment_symbols.items():
+                try:
+                    data = yf.download(symbol, period="1mo", progress=False)
+                    if (
+                        len(data) > 0 and "Close" in data.columns
+                    ):  # Series比較エラー修正
+                        current_val = data["Close"].iloc[-1]
+                        avg_val = data["Close"].mean()
+                        # 平均との乖離率
+                        deviation = (current_val / avg_val - 1) * 100
+                        sentiment_data[name] = float(deviation)
+                    else:
+                        sentiment_data[name] = 0.0
+                except Exception as symbol_error:
+                    self.logger.warning(
+                        f"センチメント指標取得エラー {name}: {symbol_error}"
+                    )
+                    sentiment_data[name] = 0.0
+            return sentiment_data
+        except Exception as e:
+            self.logger.error(f"センチメント指標取得エラー: {e}")
+            return {}
