+++ models_refactored/precision/precision_87_system.py
@@ -36,7 +36,7 @@ class Precision87BreakthroughSystem(BaseStockPredictor):
         if self._meta_learner is None:
             try:
                 # MetaLearningOptimizer は元のml_modelsにまだ存在するためそのまま維持
-                from models.ml_models import MetaLearningOptimizer
+                from models.meta_learning import MetaLearningOptimizer
 
                 self._meta_learner = MetaLearningOptimizer()
             except ImportError:
@@ -50,7 +50,7 @@ class Precision87BreakthroughSystem(BaseStockPredictor):
         if self._dqn_agent is None:
             try:
                 # DQNReinforcementLearner は元のml_modelsにまだ存在するためそのまま維持
-                from models.ml_models import DQNReinforcementLearner
+                from models.deep_learning import DQNReinforcementLearner
 
                 self._dqn_agent = DQNReinforcementLearner()
             except ImportError:
